removed old navbar and old map
