--- components/map2.tsx
@@ -1,1843 +1,0 @@
-import { useEffect, useRef, useLayoutEffect } from "react";
-import styles from "./MySVGComponent.module.css";
-
-interface Map2Props {
-  clickedCountryId: string | null;
-  setClickedCountryId: (id: string | null) => void;
-}
-
-const Map2: React.FC<Map2Props> = ({
-  clickedCountryId,
-  setClickedCountryId,
-}) => {
-  const contRef = useRef<HTMLDivElement | null>(null); //ref to parent cont
-  const clickedCountryIdRef = useRef<string | null>(null); //ref to clicked country id
-
-  useLayoutEffect(() => {
-    clickedCountryIdRef.current = clickedCountryId;
-    console.log("current clicked country id is %s", clickedCountryId);
-    console.log(
-      "current clicked country id ref is %s",
-      clickedCountryIdRef.current
-    );
-  }, [clickedCountryId]);
-
-  useEffect(() => {
-    const contElement = contRef.current;
-    if (!contElement) {
-      console.log("err: container element of Map2 component not found");
-      return;
-    }
-
-    const svgElement = contElement.querySelector("svg");
-    if (!svgElement) {
-      console.log("err: could not get countries");
-      return;
-    }
-
-    const countries = svgElement.querySelectorAll<SVGElement>(".country");
-
-    const mouseoverHandler = (event: Event) => {
-      const target = event.target as SVGElement;
-      const parent = target.parentElement;
-      if (parent) {
-        parent.style.fill = "#FFE5B4";
-      } else {
-        console.log("No parent found!");
-      }
-    };
-
-    const mouseoutHandler = (event: Event) => {
-      const target = event.target as SVGElement;
-      const parent = target.parentElement;
-      if (parent) {
-        parent.style.fill =
-          clickedCountryIdRef.current === parent.id ? "#FF0000" : "#005500";
-      } else {
-        console.log("No parent found");
-      }
-    };
-
-    const clickHandler = (event: Event) => {
-      const target = event.target as SVGElement;
-      const parent = target.parentElement;
-      if (parent) {
-        console.log("%s clicked", parent.id);
-        parent.style.fill = "#FF0000";
-        setClickedCountryId(parent.id);
-      } else {
-        console.log("err: can't find parent");
-      }
-    };
-
-    countries.forEach((country) => {
-      country.addEventListener("mouseover", mouseoverHandler);
-      country.addEventListener("mouseout", mouseoutHandler);
-      country.addEventListener("click", clickHandler);
-    });
-
-    // Cleanup function
-    return () => {
-      countries.forEach((country) => {
-        country.removeEventListener("mouseover", mouseoverHandler);
-        country.removeEventListener("mouseout", mouseoutHandler);
-        country.removeEventListener("click", clickHandler);
-      });
-    };
-  });
-
-  return (
-    <>
-      <div ref={contRef} className={styles.mapContainer}>
-        <svg
-          className={styles.mapSvg}
-          xmlns="http://www.w3.org/2000/svg"
-          data-xmlns-dc="http://purl.org/dc/elements/1.1/"
-          data-xmlns-cc="http://creativecommons.org/ns#"
-          data-xmlns-rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#"
-          data-xmlns-sodipodi="http://sodipodi.sourceforge.net/DTD/sodipodi-0.dtd"
-          data-xmlns-inkscape="http://www.inkscape.org/namespaces/inkscape"
-          version="1.0"
-          viewBox="0 0 2502.5147 1517.4011"
-          preserveAspectRatio="xMidYMid meet"
-          id="svg1926"
-          data-inkscape-version="0.91 r13725"
-          data-sodipodi-docname="Southeast Asia outline.svg"
-          data-inkscape-export-filename="C:\Users\Paul\Pictures\Image-WAGGGSAsiaPacificMap-World_notpink.png"
-          data-inkscape-export-xdpi="90"
-          data-inkscape-export-ydpi="90"
-        >
-          <metadata id="metadata2476">
-            <g data-rdf-rdf>
-              <g data-cc-work data-rdf-about="">
-                <g data-dc-format="image/svg+xml" />
-                <g
-                  data-dc-type
-                  data-rdf-resource="http://purl.org/dc/dcmitype/StillImage"
-                />
-                <g data-dc-title />
-              </g>
-            </g>
-          </metadata>
-          <g
-            id="namedview2474"
-            data-sodipodi-namedview
-            data-pagecolor="#ffffff"
-            data-bordercolor="#666666"
-            data-borderopacity="1"
-            data-objecttolerance="10"
-            data-gridtolerance="10"
-            data-guidetolerance="10"
-            data-inkscape-pageopacity="0"
-            data-inkscape-pageshadow="2"
-            data-inkscape-window-width="1920"
-            data-inkscape-window-height="1001"
-            data-showgrid="false"
-            data-fit-margin-top="0"
-            data-fit-margin-left="0"
-            data-fit-margin-right="0"
-            data-fit-margin-bottom="0"
-            data-inkscape-zoom="0.64"
-            data-inkscape-cx="711.65077"
-            data-inkscape-cy="363.49322"
-            data-inkscape-window-x="-9"
-            data-inkscape-window-y="-9"
-            data-inkscape-window-maximized="1"
-            data-inkscape-current-layer="svg1926"
-          />
-
-          <defs id="defs1930" />
-          <style type="text/css" id="style_css_sheet" />
-          <g
-            transform="matrix(6.4876122,0,0,6.4876122,-12925.057,-3966.0743)"
-            style={{
-              fill: clickedCountryId === "Indonesia" ? "#FF0000" : "#005500",
-              fillOpacity: 0.98823529,
-              stroke: "#ffffff",
-              strokeWidth: 0.5257113,
-              strokeMiterlimit: 3.98288536,
-              strokeDasharray: "none",
-              strokeOpacity: 1,
-            }}
-            id="Indonesia"
-            className="country"
-          >
-            <path
-              d="m 2324.809,838.583 c -0.508,0.321 -0.365,0.561 0.43,0.72 -0.144,-0.24 -0.287,-0.48 -0.43,-0.72"
-              id="path2528"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2326.318,819.793 c -0.652,0.219 -0.726,0.962 -0.5,1.51 0.039,-0.546 0.206,-1.049 0.5,-1.51"
-              id="path2530"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2327.479,822.093 c -0.534,-0.118 -0.584,0.122 -0.15,0.721 0.049,-0.241 0.1,-0.481 0.15,-0.721"
-              id="path2532"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2331.358,812.813 c -0.908,0.194 -0.762,1.996 0.216,1.72 0.574,-0.161 0,-1.467 -0.216,-1.72"
-              id="path2540"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2333.088,817.413 c -0.023,0.146 -0.046,0.293 -0.069,0.439 0.515,-0.035 1.019,-0.131 1.51,-0.289 -0.481,-0.044 -0.962,-0.094 -1.441,-0.15"
-              id="path2542"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2335.038,810.213 c 0,1.131 1.9,2.56 2.722,1.306 0.586,-0.893 -2.272,-1.262 -2.722,-1.306"
-              id="path2544"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2335.898,818.493 c -0.047,0.24 -0.094,0.48 -0.141,0.72 1.857,0.127 3.455,1.141 5.25,1.521 0.761,0.161 1.572,0.506 2.365,0.295 0.881,-0.234 1.877,-0.761 2.815,-0.586 -0.764,-1.121 -2.729,-1.048 -3.909,-1.19 -2.13,-0.257 -4.236,-0.539 -6.38,-0.76"
-              id="path2546"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2349.218,876.813 c -0.073,0.24 -0.146,0.48 -0.22,0.72 1.486,-0.494 3.094,-0.271 4.629,-0.279 1.505,-0.007 2.393,-0.777 3.549,-1.665 0.508,-0.39 3.661,-4.271 3.027,-5.057 -1.832,-2.268 -3.027,-2.504 -5.834,-1.333 -2.886,1.205 -4.007,4.999 -5.151,7.614"
-              id="path2548"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2358.068,877.393 c 0.243,-0.073 0.486,-0.146 0.729,-0.22 -0.373,-0.733 -0.336,-2.142 -1.543,-1.655 -1.604,0.646 -0.328,1.838 0.814,1.875"
-              id="path2550"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2357.718,862.853 c 0.034,0.568 0.394,0.939 0.931,1.08 -0.051,-0.548 -0.387,-0.96 -0.931,-1.08"
-              id="path2552"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2303.638,813.313 c 0.23,-0.566 0.134,-1.07 -0.29,-1.511 2.246,-0.571 4.11,-0.697 5.859,-2.308 1.623,-1.496 3.313,-2.133 5.495,-1.595 2.206,0.544 3.307,2.173 5.206,2.992 1.922,0.828 3.902,-0.273 5.689,1.121 -1.366,1.014 -0.053,2.375 0.531,3.397 0.789,1.381 -0.11,2.533 -0.619,3.807 -1.302,3.264 0.638,7.164 2.108,10.006 0.504,-0.734 -0.187,-2.384 0.64,-2.928 1.164,-0.767 1.02,2.666 1.031,2.992 0.024,0.653 0.24,0.813 0.686,1.25 0.641,0.627 0.34,1.473 0.909,2.113 1.103,1.238 3.365,1.9 4.636,0.582 0.924,-0.958 1.875,-1.807 2.821,-2.738 0.994,-0.979 1.185,-2.29 2.513,-3.13 1.19,-0.753 0.848,-3.544 2.365,-3.553 1.798,-0.011 2.664,0.159 4.346,-0.718 1.205,-0.628 1.262,-1.049 0.725,-2.175 -0.412,-0.867 2.227,-2.095 2.76,-2.378 1.34,-0.713 2.581,-1.206 3.999,-0.475 1.851,0.955 4.059,1.622 5.718,2.872 1.394,1.05 2.979,1.629 4.53,2.414 1.061,0.537 3.144,2.179 4.39,1.743 1.572,-0.551 2.229,0.647 3.839,0.625 1.69,-0.023 2.529,1.723 4.204,1.496 -0.209,5.034 -0.604,10.057 -0.946,15.083 -0.358,5.262 -0.126,10.67 -0.67,15.9 -0.174,1.674 -1.446,2.87 -1.373,4.648 0.03,0.733 0.979,1.129 0.979,1.589 0,1.117 -0.19,2.264 -0.267,3.38 -0.361,5.254 -0.761,10.525 -1.253,15.77 -2.155,-1.087 -2.599,-3.551 -4.237,-5.113 -0.81,-0.772 -3.899,-2.735 -2.453,-4.097 -0.943,0.738 -2.625,0.425 -3.754,0.874 -0.961,0.382 -2.739,0.47 -2.226,-1.093 -0.479,0.587 -1.705,1.694 -2.521,1.702 -1.234,0.012 0.207,-2.112 0.335,-2.412 0.418,-0.979 1.788,-3.176 0.792,-3.71 -0.898,-0.481 -1.156,-1.665 -1.847,-2.351 1.177,-1.015 2.737,0.76 3.67,-0.79 -1.626,1.623 -3.75,-0.935 -4.89,-1.939 1.506,-1.73 3.224,0.289 4.75,0.359 -1.328,-0.516 -2.456,-1.435 -3.67,-2.16 1.417,-0.249 0.197,-0.685 -0.271,-1.174 -0.665,-0.692 -1.319,-1.525 -1.586,-2.465 -0.344,-1.209 -1.464,-3.473 0.127,-4.071 -1.57,0.58 -1.663,-1.115 -0.22,-1.37 -0.538,-10e-4 -1.882,0.854 -2.3,0.221 -0.501,-0.757 0.504,-1.75 0.14,-2.59 -0.462,1.846 -1.827,-1.1 -2.429,-1.36 -1.027,-0.444 -1.088,-0.699 -1.872,-1.059 -0.606,-0.278 -2.067,-0.689 -2.431,-1.298 -0.501,-0.84 -3.599,-0.908 -4.561,-1.492 -1.51,-0.916 -3.265,-1.914 -4.954,-2.458 -1.63,-0.524 -4.032,-0.04 -5.454,-0.741 -1.27,-0.625 -2.959,-1.49 -3.495,-2.9 -0.498,-1.31 1.593,-1.104 2.306,-1.223 -0.87,-0.263 -1.965,-0.708 -2.662,0.173 -0.785,0.992 -1.303,-0.535 -2.019,-0.753 -0.011,1.476 -2.313,0.391 -1.229,-0.64 -1.574,0.015 -2.189,0.032 -2.59,-1.881 -0.07,0.556 -0.371,1.01 -0.822,0.439 -0.637,-0.807 -0.133,-1.311 -0.207,-2.154 -0.135,-1.543 1.401,-2.415 1.459,-3.975 -0.381,0.81 -1.73,1.097 -1.73,2.107 0,0.902 0.079,2.304 -1.22,2.223 1.127,0.504 0.233,1.092 -0.14,1.72 -0.253,0.227 -0.492,0.467 -0.72,0.72 -0.039,0.479 -0.065,0.959 -0.08,1.44 -0.351,1.022 -2.431,2.102 -3.514,1.8 -0.959,-0.268 -1.52,-2.512 -1.553,-3.351 -0.022,-0.566 0.728,-0.619 1.116,-0.799 1.027,-0.478 -0.55,-1.761 -0.29,-2.61 -1.239,0.747 -2.776,-3.252 -4.513,-3.322 -2.692,-0.11 -1.245,-1.962 0.692,-2.158 2.014,-0.205 3.465,1.791 4.861,-0.546 1.046,-1.749 4.242,-2.306 3.419,0.546 0.222,-0.485 0.389,-0.988 0.5,-1.51 0.286,1.381 1.871,-0.327 1.81,1.15 0.216,-1.199 0.629,-0.532 0.931,0.08 0.291,-0.627 0.341,-1.44 -0.221,-1.95 0.498,0.105 0.981,0.059 1.45,-0.14 -0.416,-0.501 -0.393,-0.981 0.07,-1.44 -1.54,1.146 -0.354,-0.82 -0.07,-1.15 -1.65,0.138 -2.539,1.187 -4.312,0.912 -1.105,-0.171 -2.58,0.874 -3.533,0.523 -1.173,-0.431 -1.666,0.369 -2.586,-0.545 -0.318,-0.316 -1.225,0.263 -1.531,0.282 -1.019,0.063 -1.828,-0.866 -2.588,-1.392 0.188,-0.245 0.381,-0.485 0.58,-0.72 -0.812,0.142 -1.544,-0.808 -1.254,-1.58 0.108,-0.288 0.497,-0.682 0.441,-1.015 -0.14,-0.848 -1.463,0.532 -0.558,-1.005 -0.736,0.144 -0.921,0.154 -0.86,-0.649 -1.039,1.507 -3.057,-1.128 -3.96,-1.08 0.86,1.133 -1.2,0.96 -1.729,0.595 -1.466,-1.016 1.384,-2.892 1.652,-3.906"
-              id="path2562"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2036.309,784.08 c 0.314,0.219 1.451,0.339 1.274,-0.423 -0.085,-0.368 -1.332,-1.134 -1.628,-1.432 -0.37,-0.372 -0.813,-0.71 -1.377,-0.593 0.287,-0.8 -0.811,-1.405 -1.37,-1.656 -0.293,-0.132 -0.533,-0.73 -0.925,-0.408 -0.726,0.596 -0.711,1.198 0.012,1.826 1.205,1.048 3.323,1.058 4.014,2.686"
-              id="path2206"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2211.628,849.383 c -0.336,1.213 0.309,2.326 0.141,3.53 1.116,-0.52 1.765,1.628 3.014,0.699 0.854,-0.635 0.867,-0.757 1.971,-0.63 1.038,0.12 1.731,-1.067 2.726,-0.29 -0.095,-1.688 -1.434,-3.145 -0.813,-4.877 0.757,-2.118 1.158,-3.503 0.897,-5.824 -0.219,-1.941 0.217,-3.465 0.394,-5.353 0.081,-0.862 -0.01,-1.718 0.087,-2.582 0.157,-1.448 -0.496,-1.446 -0.854,-2.674 -0.696,-2.388 3.441,-4.34 5.248,-3.774 1.632,0.511 1.127,4.159 0.268,5.113 -2.415,2.681 1.67,5.731 3.978,6.971 0.979,0.526 -1.481,4.198 -0.484,5.693 0.538,0.806 2.31,1.219 3.218,1.352 1.117,0.163 0.835,-1.646 0.958,-2.285 0.278,-1.438 2.311,-1.176 3.322,-1.587 0.601,-0.245 3.034,0.998 1.633,-0.752 1.093,1.851 2.566,-1.901 0.65,-2.45 -0.344,1.761 -0.814,-0.665 -1.73,-0.65 0.193,-0.264 0.387,-0.527 0.58,-0.79 -0.793,-0.034 -1.777,-1.07 -2.23,-1.649 -0.738,-0.945 0.23,-1.722 -0.439,-2.67 0.357,0.088 0.717,0.161 1.08,0.22 -0.826,-0.502 -0.605,-1.718 0.439,-1.66 -0.192,-0.369 -0.445,-0.692 -0.757,-0.967 -0.572,-0.628 -0.157,-1.298 -0.543,-1.983 -0.323,-0.574 -1.08,-0.618 -1.439,-1.149 -0.408,-0.604 -0.542,-1.566 -0.877,-2.239 -0.505,-1.015 -0.93,-2.009 -2.001,-2.533 -0.873,-0.427 -1.288,-2.255 -2.452,-1.708 0.053,-0.622 0.076,-1.246 0.069,-1.87 1.216,1.032 2.179,1.974 3.412,0.272 1.069,-1.477 2.184,-1.498 3.695,-2.342 2.074,-1.158 3.592,-6.019 6.311,-5.933 0.776,0.025 0.915,2.268 1.813,1.323 0.536,-0.563 1.557,-2.003 1.012,-2.805 -0.906,-1.332 -4.429,-0.923 -5.512,-0.026 0.588,0.062 1.092,0.302 1.51,0.72 -1.418,0.05 -5.495,-0.944 -5.992,0.892 -0.333,1.229 -4.032,1.183 -3.798,-0.531 -1.233,0.756 -2.511,2.01 -3.132,3.328 -0.675,1.435 -0.985,2.282 -2.646,1.92 -1.088,-0.236 -2.12,-0.52 -2.419,-1.748 -0.151,-0.619 -0.208,-1.245 -0.549,-1.804 -0.628,-1.028 -1.056,-0.191 -1.751,-0.697 -2.237,-1.626 -2.334,-6.686 -1.366,-8.934 0.555,-1.289 1.482,-2.734 2.756,-3.417 1.052,-0.563 3.631,0.367 4.729,0.766 0.779,0.282 2.271,-0.37 2.969,-0.692 1.44,-0.667 1.839,0.353 3.234,0.565 1.552,0.235 3.62,-0.605 5.28,-0.564 1.987,0.05 3.264,0.162 4.854,1.31 1.977,1.428 8.121,0.465 9.731,-1.418 1.106,-1.293 1.495,-2.742 2.826,-3.938 1.433,-1.288 1.367,-3.47 2.913,-4.622 -0.544,-0.344 -0.854,-0.944 -0.65,-1.584 -0.566,0.146 -1.361,-0.254 -1.838,0.151 -0.717,0.607 -0.676,1.793 -1.658,2.094 -1.499,0.459 -0.546,0.921 -1.254,1.859 -0.408,0.541 -1.41,0.332 -1.696,1.073 -0.717,1.856 -3.561,2.187 -5.139,1.846 -1.337,-0.289 -2.918,-0.597 -4.292,-0.528 -0.772,0.039 -1.511,0.894 -2.146,0.638 -1.99,-0.802 -3.397,-1.183 -5.531,-1.405 -1.097,-0.114 -1.736,-0.437 -2.902,-0.319 -1.33,0.134 -1.728,-1.68 -2.401,-1.667 -0.869,0.016 -4.097,-1.573 -4.282,0.233 -0.252,2.461 -2.484,4.805 -4.289,2.223 -0.172,0.807 -0.435,1.42 -1.06,1.971 -0.845,0.745 -0.673,1.896 -1.494,2.559 -1.567,1.263 0.478,3.463 -1.127,4.61 -0.133,-0.566 -0.539,-0.95 -1.149,-0.79 0.396,1.528 1.579,1.154 1.095,2.916 -0.441,1.606 1.026,3.078 0.565,4.715 -0.333,-0.61 -0.472,-1.314 -1.011,-1.801 -0.9,2.437 -2.817,3.71 -2.95,6.48 -0.06,1.237 0.049,2.556 -0.04,3.761 -0.104,1.412 -1.424,2.327 -1.52,3.838 -0.126,2.001 -3.397,2.924 -2.887,4.595 0.287,0.941 1.041,0.353 0.229,1.686 -0.702,1.154 0.221,3.653 0.688,4.841 0.649,-0.73 2.326,-1 3.235,-1.049 1.989,-0.107 0.412,1.822 1.059,2.978 1.073,1.92 0.815,3.562 0.475,5.773 -0.285,1.81 -0.747,4.628 -1.964,6.046"
-              id="path2348"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2231.498,848.593 c -2.326,0.806 -0.24,3.407 -0.117,3.295 0.623,-0.57 0.563,-2.649 0.117,-3.295"
-              id="path2366"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2232.148,808.703 c -0.389,-0.723 -1.41,-0.752 -1.44,0.22 0.521,0.197 1.001,0.123 1.44,-0.22"
-              id="path2368"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2233.229,807.913 c -0.46,-0.655 -1.03,-0.636 -1.511,0 0.47,0.332 1.032,0.32 1.511,0"
-              id="path2370"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2235.678,851.692 c 0.298,-0.37 1.039,-1.164 0.767,-1.706 -0.496,-0.992 0.499,-1.438 0.767,-2.293 0.146,-0.467 0.752,-3.654 -0.532,-2.964 -0.635,0.341 -1.474,0.79 -2.024,1.247 -1.035,0.858 0.53,1.957 -0.207,2.874 -0.284,0.354 -1.027,1.576 -0.789,2.043 0.387,0.755 1.656,0.531 1.65,-0.359 0.111,0.389 0.234,0.776 0.368,1.158"
-              id="path2372"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2240.138,842.763 c -1.85,0.034 -2.527,4.02 -2.61,5.432 -0.086,1.473 -0.663,2.108 -1.204,3.452 -0.671,1.665 0.364,1.221 1.295,2.206 0.448,-0.613 0.496,-1.433 0.968,-2.019 0.476,-0.59 1.353,-0.269 1.805,-0.833 0.648,-0.806 -0.689,-2.225 -1.482,-1.399 0.217,-0.876 0.173,-1.794 0.416,-2.667 0.16,-0.575 1.226,-2.482 1.464,-0.793 0.803,-1.335 -0.734,-2.109 -0.652,-3.379"
-              id="path2378"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2242.229,818.853 c -0.033,-0.511 0.014,-1.014 0.14,-1.51 2.749,1.425 2.058,-4.005 -0.14,-0.65 -0.717,-1.063 0.721,-1.362 -0.734,-1.81 -0.498,-0.153 -1.085,-0.216 -1.547,0.084 -1.056,0.685 -0.807,2.448 -0.529,3.456 0.808,-0.414 1.267,-1.228 1.65,-2.01 0.79,0.568 0.035,2.357 1.16,2.44"
-              id="path2386"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2240.068,839.083 c -0.821,0.344 -0.724,1.849 0.083,2.164 1.204,0.47 1.9,-1.905 -0.083,-2.164"
-              id="path2388"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2242.878,815.043 c -0.239,0.305 -0.393,0.232 -0.149,0.72 0.049,-0.24 0.099,-0.48 0.149,-0.72"
-              id="path2394"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2243.229,819.293 c -0.317,0.124 -0.484,0.364 -0.5,0.72 0.167,-0.24 0.333,-0.48 0.5,-0.72"
-              id="path2396"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2243.309,850.103 c 0.074,0.512 10e-4,0.716 0.649,0.721 -0.217,-0.241 -0.433,-0.481 -0.649,-0.721"
-              id="path2398"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2244.528,851.973 c 0.056,0.407 0.295,0.624 0.72,0.65 -0.24,-0.217 -0.479,-0.434 -0.72,-0.65"
-              id="path2400"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2246.548,855.653 c -0.517,0.518 -0.35,0.758 0.5,0.72 -0.167,-0.24 -0.333,-0.48 -0.5,-0.72"
-              id="path2404"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2251.229,819.003 c -1.26,1.836 -0.257,3.436 1.939,2.73 0.348,-0.215 0.707,-0.407 1.08,-0.575 0.418,0.084 0.84,0.125 1.267,0.122 0.796,-0.399 2.186,-0.082 2.554,-1.126 -0.576,0.07 -0.959,-0.232 -1.08,-0.79 -0.93,1.312 -4.37,-0.854 -5.76,-0.361"
-              id="path2408"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2258.788,780.84 c -0.741,-0.173 -0.815,0.067 -0.22,0.72 0.074,-0.24 0.147,-0.48 0.22,-0.72"
-              id="path2414"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2260.298,775.656 c 0.432,-0.902 -0.837,-3.777 -1.95,-2.521 1.179,0.324 0.781,2.031 1.95,2.521"
-              id="path2416"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2258.788,820.083 c -1.035,1.085 1.618,1.141 2.134,1.149 1.403,0.023 3.48,0.24 4.557,-0.859 -2.206,-0.325 -4.544,0.492 -6.691,-0.29"
-              id="path2418"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2262.888,821.733 c -1.543,0.334 -0.139,3.851 0.65,4.399 0.319,-0.558 -0.286,-3.834 -0.65,-4.399"
-              id="path2422"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2267.928,837.793 c 1.057,-0.526 3.374,-0.876 4.047,-1.885 0.852,-1.276 0.384,-2.212 -0.946,-2.505 1.056,-1.191 -1.239,-1.925 -2.031,-2.122 -1.488,-0.37 -2.813,0.093 -4.289,0.253 -1.807,0.197 -2.184,1.092 -1.37,2.834 0.854,1.831 2.915,2.529 4.589,3.425"
-              id="path2424"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2268.068,771.84 c 0.197,-0.577 -0.021,-1.155 -0.5,-1.512 0.143,0.512 0.309,1.016 0.5,1.512"
-              id="path2426"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2268.358,770.184 c 0.786,-0.813 0.977,-2.084 0.828,-3.163 -0.076,-0.555 -0.561,-2.199 -1.073,-0.882 -0.356,0.917 0.707,1.644 0.359,2.339 -0.318,0.639 -0.96,1.113 -0.114,1.706"
-              id="path2428"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2272.898,806.973 c -0.902,-0.387 -1.618,1.641 -1.011,2.16 1.051,0.898 1.153,-1.614 1.011,-2.16"
-              id="path2432"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2272.898,809.933 c -0.867,0.451 -0.427,1.808 0.58,1.51 -0.109,-0.536 -0.302,-1.04 -0.58,-1.51"
-              id="path2434"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2273.618,797.253 c -0.218,0.343 -0.09,0.476 0.22,0.79 -0.073,-0.263 -0.147,-0.527 -0.22,-0.79"
-              id="path2438"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2275.128,807.333 c -0.237,0.315 -0.478,0.628 -0.72,0.939 -0.244,-0.235 -0.484,-0.476 -0.72,-0.72 -0.662,1.998 0.313,1.269 0.939,2.66 0.14,0.308 -0.785,1.63 0.2,1.393 0.534,-0.128 1.11,-0.434 1.305,0.269 0.224,0.811 3.097,-0.549 0.697,-1.21 -0.98,-0.27 -1.211,-0.835 -0.682,-1.684 0.332,-0.528 -0.724,-1.353 -1.019,-1.647"
-              id="path2440"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2274.118,801.433 c -0.183,0.34 -0.33,0.251 0,0.721 0,-0.241 0,-0.481 0,-0.721"
-              id="path2442"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2274.479,817.053 c -1.113,0.865 -0.533,2.422 0.79,2.601 0.559,0.075 1.053,-0.36 1.606,-0.36 0.601,0 1.2,0.191 1.802,0.225 2.052,0.115 0.038,-1.854 -0.576,-2.167 -1.345,-0.687 -2.296,-1.338 -3.622,-0.299"
-              id="path2444"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2274.408,798.263 c -0.167,-0.023 -0.333,-0.047 -0.5,-0.07 0.13,0.287 0.297,0.551 0.5,0.791 0,-0.241 0,-0.481 0,-0.721"
-              id="path2446"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2276.278,804.603 c 0.061,0.894 -0.289,1.828 0.341,2.603 0.609,0.749 1.257,1.11 1.612,2.065 0.389,1.043 2.321,3.292 3.236,2.742 -2.604,-2.668 -2.908,-5.887 -3.423,-9.319 -0.224,-1.487 -0.164,-2.71 1.688,-1.766 1.575,0.803 3.86,1.29 5.614,1.445 -1.102,-0.539 -1.122,-1.883 -2.016,-2.589 -0.896,-0.71 -3.43,-0.594 -2.333,-2.428 0.732,-1.224 2.666,-1.263 3.032,-2.787 0.248,-1.028 -0.479,-2.753 -0.043,-3.506 -1.845,-0.469 -3.863,1.355 -4.61,2.881 1.723,0.083 0.041,0.977 -0.438,1.166 -1.042,0.411 -0.587,1.532 -1.079,2.297 -0.511,0.793 -1.592,-0.021 -1.913,-0.549 -0.573,-0.939 0.617,-1.547 1.255,-1.967 1.316,-0.866 2.176,-3.534 1.455,-5.052 -0.309,-0.65 -1.402,-0.886 -1.137,-1.757 0.254,-0.835 1.391,-1.351 1.066,-2.346 -1.76,0.013 -2.56,2.235 -2.98,3.607 -0.446,1.454 -1.601,4.025 -1.339,5.464 0.128,0.707 0.433,1.445 0.83,2.042 0.159,0.284 0.386,0.492 0.68,0.623 -0.479,0.6 -0.65,1.335 -0.575,2.089 0.097,0.986 0.438,1.55 1.025,2.284 0.591,0.737 0.183,1.935 0.052,2.758"
-              id="path2452"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2275.348,814.823 c -0.263,0.193 -0.526,0.387 -0.789,0.58 0.535,0.071 1.039,-0.025 1.51,-0.29 -0.241,-0.096 -0.481,-0.193 -0.721,-0.29"
-              id="path2454"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2274.908,832.893 c -0.434,0.602 -0.24,0.842 0.58,0.721 -0.193,-0.241 -0.386,-0.481 -0.58,-0.721"
-              id="path2456"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2275.988,831.963 c -0.174,0.26 -0.244,0.547 -0.21,0.86 0.272,0.003 0.536,-0.044 0.79,-0.141 -0.168,-0.26 -0.361,-0.499 -0.58,-0.719"
-              id="path2458"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2277.868,829.873 c -0.228,0.369 -0.467,0.729 -0.72,1.08 0.569,-0.119 0.851,-0.502 0.72,-1.08"
-              id="path2462"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2293.348,834.692 c 1.608,0.93 3.256,1.576 4.729,2.727 1.714,1.338 1.954,-0.668 1.676,-2.005 -0.172,-0.823 -1.277,-1.088 -1.44,-1.871 -0.167,-0.808 -0.167,-1.71 -0.968,-2.188 -0.74,-0.442 -1.02,-1.019 -1.987,-0.832 -0.971,0.188 -1.902,0.053 -2.692,-0.572 -0.769,-0.609 -1.755,-1.071 -2.755,-1.04 -1,0.031 -1.467,0.844 -2.352,1.038 -0.781,0.17 -1.081,-0.237 -1.496,-0.725 -0.147,-0.174 -2.596,0.128 -2.949,0.103 -3.365,-0.244 -6.913,2.145 -5.535,5.877 0.027,-0.76 1.074,-3.848 2.16,-3.601 -0.345,1.016 1.459,2.639 2.324,2.8 0.355,0.066 4.403,-3.298 3.436,-0.71 0.633,-0.139 4.364,1.362 4.576,0.075 0.183,-1.107 3.08,-0.188 3.273,0.924"
-              id="path2464"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2278.298,835.703 c -0.572,0.274 -1.009,0.843 -0.87,1.51 0.85,-0.107 1.439,-0.767 2.24,-1 -0.771,1.292 1.033,0.055 0.86,-1.16 -0.811,0.056 -1.376,0.822 -2.23,0.65"
-              id="path2466"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2279.448,784.8 c 0.218,-0.328 0.376,-0.222 0.07,-0.72 -0.023,0.24 -0.046,0.48 -0.07,0.72"
-              id="path2468"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2280.598,787.176 c 2.707,-0.609 3.705,-2.908 2.091,-5.328 -0.318,0.781 -1.199,0.979 -1.731,1.558 -0.921,1 -0.943,2.623 -0.36,3.77"
-              id="path2470"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2281.468,835.053 c -0.211,0.268 -0.238,0.558 -0.08,0.87 0.303,-0.144 0.569,-0.337 0.801,-0.58 -0.241,-0.097 -0.481,-0.194 -0.721,-0.29"
-              id="path2474"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2282.548,834.843 c -0.267,0.747 0.321,1.383 1.08,1.01 -0.088,-0.568 -0.491,-0.986 -1.08,-1.01"
-              id="path2478"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2288.958,804.603 c 0.376,0.646 0.943,1.178 1.65,1.44 -0.5,-0.517 -0.997,-1.107 -1.65,-1.44"
-              id="path2482"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2293.778,822.093 c 0.492,0.166 0.995,0.286 1.51,0.36 0.553,-0.013 0.848,-0.514 1.44,-0.431 -0.581,-0.797 0.632,-1.185 0.01,-2.395 -0.67,-1.297 -3.542,0.846 -4.687,1.026 0.178,0.811 0.903,1.41 1.731,1.44"
-              id="path2484"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2292.988,814.533 c -0.299,0.148 -0.565,0.342 -0.8,0.58 0.69,0.382 0.957,0.189 0.8,-0.58"
-              id="path2486"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2294.788,804.673 c 0.12,0.24 0.24,0.48 0.36,0.72 0.096,-0.602 -0.089,-0.413 -0.36,-0.72"
-              id="path2488"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2299.468,808.413 c 3.009,0.851 0,-2.043 -0.649,-2.73 0.849,-0.12 1.418,0.791 1.875,1.338 0.854,1.021 1.083,0.142 1.911,0.556 1.567,0.783 1.708,-1.458 0.567,-2.083 -1.075,-0.588 -6.687,-1.43 -6.804,0.688 1.433,-0.965 0.646,0.594 -0.22,0.29 0.688,0.882 2.255,0.28 1.949,1.87 0.084,-0.544 0.477,-0.956 1.011,-1.08 0.14,0.378 0.259,0.761 0.36,1.151"
-              id="path2490"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2297.378,811.803 c -0.259,1.356 3.388,0.878 3.021,-0.51 -0.772,0.862 -2.062,0.21 -3.021,0.51"
-              id="path2494"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2298.818,808.272 c -0.43,0.178 -0.766,0.464 -1.01,0.86 0.626,0.38 2.028,-0.437 1.01,-0.86"
-              id="path2496"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2300.838,812.373 c -2.734,0.456 -1.658,2.364 -0.176,3.83 0.861,0.851 2.565,-3.366 0.176,-3.83"
-              id="path2500"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2302.638,813.243 c -0.225,0.32 -0.451,0.18 -0.29,0.72 0.096,-0.24 0.193,-0.48 0.29,-0.72"
-              id="path2506"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2303.068,839.883 c 0.117,0.447 0.104,0.633 0.649,0.72 -0.216,-0.24 -0.432,-0.48 -0.649,-0.72"
-              id="path2508"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2303.998,838.942 c -0.624,-0.094 -0.551,0.146 0.22,0.721 -0.073,-0.24 -0.147,-0.48 -0.22,-0.721"
-              id="path2510"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2318.548,840.173 c -0.29,0.997 1.68,1.209 2.16,1.72 -0.56,-0.783 -1.599,-0.939 -2.16,-1.72"
-              id="path2522"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2240.928,891.143 c 0.395,0.295 0.342,0.604 0.94,0.36 0.198,1.067 -1.772,1.406 -1.95,2.811 0.861,-0.266 1.733,-0.033 2.604,-0.199 1.135,-0.217 1.98,-1.431 3.132,-1.464 1.851,-0.053 2.484,-0.583 3.717,-1.958 0.795,-0.887 2.127,-2.009 2.461,-3.122 0.508,-1.692 -0.5192,-0.33315 0.9658,-0.82915 l 1.1162,-3.82833 -1.6238,-1.26506 c -2.979,1.204 -0.6952,1.88079 -4.2002,2.56079 -0.8735,0.16925 -0.1953,2.64953 -0.8122,3.0415 -0.6169,0.39197 -0.7788,-0.92937 -1.2076,-0.36337 -0.4288,0.566 -1.8745,-1.44941 -2.1836,-0.75807 -0.3091,0.69135 -2.7006,4.24569 -2.9586,5.01369"
-              id="path2392"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2207.958,875.013 c -0.504,0.198 -0.674,0.558 -0.51,1.08 0.671,-0.063 0.843,-0.529 0.51,-1.08"
-              id="path2346"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2103.338,861.192 c -0.306,0.134 -0.569,0.327 -0.79,0.58 0.693,0.384 0.956,0.19 0.79,-0.58"
-              id="path2278"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2114.718,857.083 c -0.639,-0.496 -3.844,-1.09 -4.47,-0.5 -0.081,-2.038 -1.974,0.629 -2.23,1.3 -0.261,0.679 -0.295,3.95 -1.729,2.66 -0.243,1.075 -0.907,1.937 -1.3,2.95 -0.426,-1.01 -1.344,-0.734 -2.08,-0.28 2.321,1.126 5.061,-0.224 7.153,1.23 0.652,0.453 1.587,0.489 2.302,0.832 1.141,0.546 -0.455,1.381 -0.624,2.17 -0.456,2.13 5.67,1.76 6.807,1.964 2.188,0.393 4.008,1.991 6.191,2.554 0.689,0.178 1.634,0.411 2.313,0.062 0.725,-0.373 0.554,-1.009 1.595,-1.039 2.654,-0.078 5.336,0.55 7.806,1.371 3.057,1.015 6.009,2.65 9.143,3.367 1.302,0.298 3.16,0.024 4.326,0.593 1.202,0.586 2.229,0.049 3.539,0.175 2.68,0.258 5.672,1.924 8.128,0.241 2.004,-1.375 4.322,1.132 6.048,2.025 0.583,0.302 4.142,0.52 4.068,0.931 -0.134,0.747 0.983,1.196 1.588,0.901 0.447,-0.219 -0.739,-1.719 -0.877,-1.931 -1.211,-1.862 0.258,-3.707 0.437,-5.623 0.112,-1.203 -1.082,-1.428 -1.951,-1.701 -1.488,-0.468 -1.949,0.181 -3.34,0.044 -1.315,-0.128 -3.104,0.826 -4.242,0.265 -0.723,-0.356 -2.795,-1.051 -3.129,-1.846 -0.286,-0.679 0.976,-3.895 -0.689,-2.483 0.328,-1.065 -0.593,-2.098 -0.51,-3.171 -0.972,0.497 -2.269,0.155 -3.311,0.255 -0.409,-0.044 -0.744,-0.227 -1.006,-0.548 -0.719,-0.623 -1.466,-0.155 -2.233,-0.566 -0.403,-0.216 -1.079,-1.891 -1.78,-1.05 -0.693,0.831 -2.086,0.688 -2.622,-0.252 -1.266,-2.224 -2.753,-2.496 -3.79,0.269 -0.35,0.931 -0.963,3.048 -2.383,2.581 -0.792,-0.261 -0.957,-0.675 -1.846,-0.614 -0.719,0.049 -1.47,0.237 -2.189,0.077 -1.408,-0.313 -2,-0.235 -3.363,-0.15 -0.789,0.05 -2.049,-0.99 -2.752,-0.608 -2.843,1.546 -3.454,-2.444 -4.539,-3.967 -1,-1.405 -0.817,-0.104 -1.721,-0.104 -1.062,0 -1.981,-1.537 -3.081,-0.987 -1.044,0.522 -1.761,-0.852 -2.404,-1.393 -1.057,-0.888 -1.852,-0.526 -3.02,-1.22 -0.213,1.01 -1.218,2.396 -2.233,1.216"
-              id="path2280"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2129.979,871.133 c 0.074,0.757 0.991,0.685 1.51,0.72 -0.461,-0.332 -0.964,-0.572 -1.51,-0.72"
-              id="path2296"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2161.729,864.293 c -0.731,-0.117 -2.604,1.334 -1.736,2.056 0.817,0.681 2.771,1.27 2.886,0.104 1.179,1.397 2.692,0.588 4.091,0.086 0.576,-0.207 1.204,-0.374 1.819,-0.237 -0.126,-0.969 0.948,-0.947 1.58,-1.149 -0.247,-1.44 -2.183,-1.103 -3.193,-0.963 -1.801,0.25 -3.682,-0.377 -5.447,0.103"
-              id="path2308"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2160.358,854.213 c -0.574,0.089 -0.691,0.502 -0.5,1 0.682,-0.01 0.835,-0.457 0.5,-1"
-              id="path2310"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2172.168,865.733 c 0.066,0.533 -0.032,0.776 0.65,0.789 -0.2,-0.276 -0.417,-0.54 -0.65,-0.789"
-              id="path2312"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2172.818,874.653 c -0.937,1.776 1.435,2.638 2.653,3.217 0.767,0.365 1.38,0.738 1.899,1.423 0.938,1.238 -0.232,0.908 -0.294,1.802 -0.04,0.585 2.322,-2.089 2.426,-2.125 0.976,-0.346 3.427,-1.315 2.2,-2.606 -0.887,-0.934 -2.646,-2.251 -3.974,-1.521 -1.518,0.837 -3.416,0.861 -4.91,-0.19"
-              id="path2314"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2173.608,866.373 c 0.208,0.183 0.447,0.183 0.72,0 -0.24,0 -0.479,0 -0.72,0"
-              id="path2316"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2179.658,863.783 c -0.605,0.324 -0.743,0.879 -0.22,1.37 0.554,-0.684 1.471,-0.788 2.239,-0.431 -0.334,-0.867 -1.237,-0.885 -2.019,-0.939"
-              id="path2318"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2180.878,880.853 c 0.229,-0.385 0.229,-0.771 0,-1.16 -0.282,0.17 -0.522,0.387 -0.72,0.65 0.24,0.17 0.48,0.34 0.72,0.51"
-              id="path2320"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2186.928,875.733 c -1.038,0.52 -2.322,1.487 -2.195,2.81 0.102,1.053 -0.264,2.756 -1.614,1.66 -1.115,1.72 4.67,1.239 4.89,1.51 -0.477,-1.053 1.043,-3.073 1.552,-3.927 0.781,-1.31 -1.631,-2.428 -2.633,-2.053"
-              id="path2322"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2198.448,881.853 c 1.548,-0.799 3.077,-1.059 3.96,-2.729 0.332,0.554 -0.424,1.702 0.145,1.998 0.611,0.317 1.038,0.084 1.602,-0.143 0.713,-0.288 1.295,0.132 1.994,0.235 -0.359,-0.479 -0.836,-0.745 -1.431,-0.8 1.375,-0.831 3.349,0.727 3.74,-1.51 -1.247,1.547 -1.093,-1.318 -1.394,-1.917 -0.727,-1.452 -2.247,0.038 -2.206,1.127 -0.214,-1.084 -1.046,-2.927 -2.281,-1.557 -0.935,1.035 -1.715,-1.003 -2.191,-1.53 -0.696,-0.771 -3.564,-0.185 -2.736,1.105 1.013,1.577 2.343,1.952 3.898,2.842 -0.489,0.656 -1.805,1.011 -2.557,1.18 -1.292,0.289 -1.445,-0.903 -1.988,-1.72 -1.267,-1.902 -5.234,-1.227 -6.554,0.281 -0.634,0.724 0.021,1.854 -0.563,2.592 -0.782,0.989 0.62,1.783 1.416,1.993 0.493,0.13 0.914,0.154 1.338,-0.174 0.785,-0.607 0.757,-0.207 1.774,-0.247 1.429,-0.057 2.596,-1.493 4.034,-1.026"
-              id="path2328"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2195.928,877.253 c 0.506,-0.612 0.796,-1.356 1.221,-2.02 -1.504,-0.688 -1.374,1.085 -1.221,2.02"
-              id="path2330"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2205.938,887.183 c 0.681,1.787 2.345,1.73 3.96,1.91 1.912,0.213 3.432,1.983 4.581,3.354 2.157,2.572 7.129,0.245 4.531,-2.63 -0.811,-0.896 -1.421,-1.641 -2.613,-1.643 0.056,0 -1.874,-2.309 -2.118,-2.528 -0.18,-0.162 -3.379,-0.111 -3.532,-0.013 -1.488,0.956 -3.557,-0.562 -4.809,1.55"
-              id="path2342"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2210.548,877.603 c -0.405,0.671 -0.903,1.455 -0.22,2.17 0.024,-0.994 1.987,-1.476 0.22,-2.17"
-              id="path2350"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2211.769,879.043 c 0.204,0.566 0.134,1.096 -0.21,1.59 0.964,0.238 0.798,-0.873 1,-1.44 -0.264,-0.05 -0.527,-0.1 -0.79,-0.15"
-              id="path2352"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2214.428,880.923 c 0.77,-0.412 2.252,0.052 3.149,-0.049 0.827,-0.093 2.071,0.088 2.884,0.254 0.889,0.182 1.361,1.176 2.382,0.711 0.705,-0.321 2.385,0.298 2.105,-1.067 2.236,0.938 3.192,0.322 5.347,-0.464 1.44,-0.525 6.486,0.188 5.743,-2.706 0.995,0.191 0.878,-1.043 1.8,-1.069 -0.409,-1.19 -1.68,-2.965 -2.02,-0.65 0.455,-0.285 0.936,-0.334 1.439,-0.149 -1.002,1.151 -2.789,1.229 -3.546,2.64 -0.915,1.708 -2.454,0.572 -3.418,-0.327 -0.914,-0.852 -2.043,0.556 -3.018,0.168 -0.515,-0.205 -0.478,0.203 -0.803,0.509 -0.509,0.481 -1.489,-0.357 -1.864,-0.628 -1.248,-0.901 -3.221,-1.463 -4.698,-1.809 -2.327,-0.544 -3.074,0.34 -5.007,1.303 -0.969,0.484 -1.6,0.471 -1.71,1.82 -0.056,0.698 0.214,2.768 1.235,1.513"
-              id="path2354"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2219.838,854.713 c -1.232,0.412 -0.631,4.976 -0.359,5.9 0.25,-1.973 0.689,-3.904 0.359,-5.9"
-              id="path2356"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2220.479,865.373 c -0.219,0.357 -0.219,0.717 0,1.08 0.277,-0.146 0.518,-0.34 0.72,-0.58 -0.241,-0.166 -0.48,-0.333 -0.72,-0.5"
-              id="path2358"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2221.708,867.533 c 0.198,0.616 1.08,0.298 1.44,0.14 -0.487,0.029 -0.967,-0.017 -1.44,-0.14"
-              id="path2360"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2224.878,896.403 c 0.253,0.106 0.492,0.056 0.72,-0.15 -0.241,0.05 -0.48,0.1 -0.72,0.15"
-              id="path2362"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2226.388,895.893 c -0.05,0.12 -0.1,0.24 -0.149,0.36 0.574,0.219 1.426,0.056 1.889,-0.344 0.232,-0.186 0.288,-0.414 0.168,-0.685 -0.487,-0.916 -1.33,0.364 -1.908,0.669"
-              id="path2364"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2234.809,897.772 c -0.308,0.466 -0.331,0.946 -0.07,1.44 1.219,-0.693 3.065,-1.898 4.39,-2.24 -0.251,-0.464 -0.467,-0.944 -0.649,-1.439 -1.371,0.628 -2.143,1.694 -3.671,2.239"
-              id="path2374"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2236.828,878.333 c 0.07,0.213 0.14,0.426 0.21,0.64 0.303,-0.532 0.713,-0.962 1.23,-1.29 -0.586,-0.021 -1.067,0.195 -1.44,0.65"
-              id="path2380"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2238.338,876.313 c -0.334,0.282 -0.5,0.642 -0.5,1.079 0.611,-0.077 2.01,0.197 2.23,-0.649 0.28,-1.066 -1.241,-0.713 -1.73,-0.43"
-              id="path2382"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2239.559,892.513 c -0.508,0.328 -0.839,0.904 -0.86,1.51 0.936,0.125 1.125,-0.824 0.86,-1.51"
-              id="path2384"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2240.718,876.243 c 0.196,0.22 0.362,0.459 0.5,0.72 -0.562,0.229 -1.453,0.643 -1.644,1.283 -0.336,1.132 1.797,0.276 2.224,0.587 -0.146,-1.358 1.934,-2.451 3.021,-2.66 -1.404,-1.546 -2.127,0.457 -3.101,0.93 0.483,-1.04 -0.253,-0.976 -1,-0.86"
-              id="path2390"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2245.108,876.813 c -0.307,0.294 -0.45,0.654 -0.431,1.079 0.617,-0.146 1.164,0.172 1.44,0.721 0.584,-0.609 1.786,-2.078 1,-2.95 -0.388,0.801 -1.244,0.93 -2.009,1.15"
-              id="path2402"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2248.988,875.013 c -0.422,0.468 -0.988,1.66 0.359,0.94 -0.749,0.333 -1.395,1.015 -1.439,1.87 1.776,-0.169 3.818,-0.024 5.479,-0.772 1.049,-0.472 0.693,-1.576 -0.345,-1.759 -0.614,-0.108 -1.151,0.15 -1.764,0.091 -0.771,-0.074 -1.497,-0.476 -2.29,-0.37"
-              id="path2406"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2257.918,875.163 c -0.559,-0.058 -0.979,0.9 -1.08,1.29 0.539,-0.239 0.94,-0.717 1.08,-1.29"
-              id="path2410"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2258.998,874.013 c 1.506,-1.257 2.969,-0.971 4.75,-0.661 0.947,0.165 1.902,-1.672 3.1,-1.649 -1.564,-1.931 -2.374,-1.242 -4.633,-0.363 -0.631,0.246 -1.072,-0.328 -1.605,-0.099 -0.975,0.418 -1.472,1.814 -1.612,2.772"
-              id="path2420"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2271.748,869.613 c -0.309,0.378 -0.455,0.812 -0.439,1.3 0.644,-0.222 0.95,-0.727 0.439,-1.3"
-              id="path2430"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2273.258,875.303 c -0.178,0.85 0.063,0.899 0.721,0.15 -0.241,-0.05 -0.481,-0.1 -0.721,-0.15"
-              id="path2436"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2274.269,874.733 c -0.321,1.255 2.06,1.676 2.3,0.279 -0.784,0.145 -1.555,-0.063 -2.3,-0.279"
-              id="path2450"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2276.638,875.663 c 0.026,0.217 0.054,0.434 0.08,0.65 0.271,-0.179 0.508,-0.396 0.71,-0.65 -0.264,0 -0.527,0 -0.79,0"
-              id="path2460"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2281.178,865.803 c -0.323,0.193 -0.49,0.48 -0.5,0.86 0.786,0.089 0.953,-0.197 0.5,-0.86"
-              id="path2472"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2282.328,875.522 c 0.229,0.489 0.613,0.706 1.15,0.65 -0.109,-0.607 -0.572,-0.875 -1.15,-0.65"
-              id="path2476"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2288.448,872.133 c -0.642,0.559 -0.308,1.929 0.629,1.912 1.785,-0.032 0.969,-2.096 -0.629,-1.912"
-              id="path2480"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2296.798,876.893 c 1.181,-0.396 2.255,-1.135 3.24,-1.88 -1.242,0.48 -2.477,0.362 -3.24,1.88"
-              id="path2492"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2298.098,869.543 c -0.128,0.619 0.112,0.642 0.721,0.07 -0.241,-0.023 -0.481,-0.047 -0.721,-0.07"
-              id="path2498"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2299.388,868.823 c -0.116,0.193 -0.233,0.387 -0.35,0.58 0.577,0.194 0.961,-0.069 1.08,-0.65 -0.243,0.023 -0.487,0.047 -0.73,0.07"
-              id="path2502"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2304.218,866.233 c -1.183,0.072 -4.527,4.052 -3.38,5.109 -2.329,-1.073 -0.706,4.782 0.472,2.077 0.521,-1.198 1.821,-1.783 2.464,-2.923 0.531,-0.943 -0.098,-3.073 1.095,-3.474 -0.181,-0.293 -0.397,-0.556 -0.651,-0.789"
-              id="path2504"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2305.008,866.233 c 0.385,0.638 1.244,0.464 1.73,1.01 0.01,-0.987 -0.989,-1.067 -1.73,-1.01"
-              id="path2512"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2313.288,853.772 c -0.181,-0.231 -0.421,-0.301 -0.72,-0.209 0.309,0.521 -0.19,2.125 0.504,2.339 1.084,0.336 0.407,-1.744 0.216,-2.13"
-              id="path2516"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2313.868,852.913 c -0.631,-0.131 -0.654,0.109 -0.07,0.72 0.023,-0.24 0.047,-0.48 0.07,-0.72"
-              id="path2518"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2313.868,856.583 c 1.296,-0.96 2.579,-4.366 3.02,-5.9 -1.087,0.384 -0.973,1.451 -1.341,2.334 -0.519,1.244 -1.292,2.213 -1.679,3.566"
-              id="path2520"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2323.588,858.022 c -0.201,0.888 -0.364,2.092 0.36,2.811 -1.447,-0.082 -1.926,4.254 0.467,3.113 3.11,-1.485 1.093,-4.488 -0.827,-5.924"
-              id="path2524"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2327.838,861.692 c -0.517,0.327 -0.581,1.002 -0.359,1.521 0.375,-0.447 0.495,-0.954 0.359,-1.521"
-              id="path2534"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2327.838,860.473 c -0.668,0.392 -0.574,0.632 0.28,0.72 -0.094,-0.24 -0.186,-0.481 -0.28,-0.72"
-              id="path2536"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2328.838,860.613 c 0.856,-2.607 -0.609,-0.843 0,0"
-              id="path2538"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2081.738,852.123 c 0.205,-0.974 -1.196,-2.093 -1.94,-1.15 0.699,0.337 0.991,1.512 1.94,1.15"
-              id="path2230"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2027.448,754.488 c 1.084,0.341 0.526,-1.038 -0.5,-0.721 0.167,0.241 0.333,0.481 0.5,0.721"
-              id="path2202"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2044.518,759.168 c -0.635,0.37 -1.455,0.925 -2.245,0.706 -0.752,-0.209 -0.962,-0.89 -1.814,-0.912 -1.424,-0.037 -2.657,0.699 -4.121,0.558 -1.723,-0.166 -2.851,-0.73 -3.974,-2.021 -0.838,-0.963 -3.164,-2.305 -4.438,-1.413 -4.702,3.296 6.704,12.52 8.51,14.933 0.688,0.92 1.618,0.956 2.669,1.148 1.737,0.318 2.292,2.257 3.084,3.534 0.609,0.981 1.836,3.441 2.886,3.884 1.826,0.771 1.118,2.544 1.624,4.171 0.476,1.532 2.056,1.149 3.232,1.966 0.749,0.519 1.481,1.077 2.22,1.611 0.907,0.657 1.832,1.332 2.615,2.139 1.1,1.134 0.048,1.258 0.383,2.517 0.402,1.509 1.054,2.963 1.465,4.488 0.435,1.615 1.591,3.301 1.476,5.005 -0.122,1.797 1.271,0.694 1.954,1.501 1.117,1.321 2.759,1.013 3.063,3.171 0.396,2.813 4.102,4.64 4.732,7.59 0.09,0.422 -0.613,0.47 -0.279,0.925 0.453,0.617 0.853,1.268 1.197,1.951 0.781,1.548 2.401,3.582 2.461,5.338 0.058,1.698 0.016,2.776 1.061,4.249 0.787,1.109 1.992,1.89 2.653,3.085 0.788,1.424 1.2,2.859 2.454,3.991 1.241,1.12 2.979,1.871 3.738,3.444 0.75,1.552 0.617,2.463 2.053,3.645 1.455,1.197 2.746,2.562 4.139,3.829 1.205,1.095 2.377,2.129 3.824,2.883 1.443,0.752 1.635,1.31 2.348,2.596 0.378,0.682 0.874,1.321 1.352,1.936 0.411,0.438 0.855,0.84 1.334,1.204 0.295,0.266 -0.06,0.559 0.313,0.824 0.266,0.189 2.661,3.308 2.674,1.658 0.01,-0.967 -0.913,-1.706 -1.012,-2.661 -0.14,-1.342 1.969,0.391 2.221,0.614 1.793,1.591 3.544,1.598 3.33,-1.419 1.118,0.967 1.872,2.932 3.46,3.24 0.951,0.185 0.916,-2.691 0.938,-3.177 0.084,-1.872 0.363,-3.732 0.502,-5.603 0.073,-0.997 0.719,-2.411 0.274,-3.367 -0.343,-0.738 -0.582,-1.159 -0.347,-1.953 0.287,-0.965 1.07,-2.258 0.946,-3.288 -0.092,-0.764 -0.95,-0.797 -0.754,-1.83 0.17,-0.897 0.72,-1.68 1.171,-2.452 0.486,-0.833 0.868,-1.188 0.684,-2.226 -0.177,-0.987 -0.989,-0.585 -1.544,-1.165 -0.756,-0.791 -1.962,-3.43 -1.511,-4.46 -0.871,-0.142 -5.054,-1.242 -5.399,-0.36 -0.13,-0.259 -0.25,-0.521 -0.36,-0.789 -0.914,1.136 -1.032,2.736 -2.38,3.529 0.396,-0.587 2.229,-2.585 0.431,-2.6 0.972,-0.395 2.131,-1.244 1.97,-2.437 -0.121,-0.898 -1.065,-0.585 -1.626,-0.924 -0.57,-0.345 -0.886,-1.827 -1.015,-2.402 -0.38,-1.688 -0.785,-3.291 -0.997,-5.021 -0.131,-1.064 -3.248,-1.178 -4.163,-0.966 -0.72,0.167 -2.095,-1.254 -2.619,-1.69 -0.888,-0.737 -0.531,-2.284 -0.91,-3.33 0.399,0.482 0.995,0.575 1.51,0.209 -0.204,-0.833 -1.009,-0.943 -1.729,-1 0.584,-0.042 1.016,-0.455 1.16,-1.01 -0.487,-0.03 -0.968,0.017 -1.44,0.141 0.848,-0.248 1.352,-1 1.974,-1.562 0.479,-0.433 1.731,-0.306 2.022,-0.792 0.48,-0.803 -1.284,-2.545 -1.744,-3.063 -1.277,-1.436 -2.226,-1.316 -3.699,-0.269 -1.14,0.809 -4.025,2.855 -5.253,1.225 2.551,1.463 7.458,-3.06 2.811,-4.109 -0.741,-0.167 -1.505,0.154 -2.23,-0.141 -0.822,-0.333 -1.168,-1.265 -1.833,-1.792 -1.366,-1.083 -0.599,-3.308 -2.092,-4.458 -0.551,-0.424 -1.123,-1.082 -1.719,-1.425 -0.634,-0.366 -1.478,-0.053 -2.18,-0.418 -1.142,-0.595 -1.051,-2.362 -1.786,-3.332 -0.361,-0.478 -2.934,-2.989 -3.498,-1.49 -0.42,1.116 0.939,2.541 1.077,3.627 -0.469,-0.479 -0.604,-1.123 -0.985,-1.655 -0.631,-0.88 -1.784,-1.012 -2.545,-1.729 -1.259,-1.187 -1.389,-3.085 -2.59,-4.319 -0.344,0.466 -0.414,0.97 -0.21,1.512 -0.203,-0.362 -1.45,-2.813 -1.01,-0.864 -0.809,-0.93 0.397,-2.088 -0.828,-2.888 -1.11,-0.724 -1.971,-1.987 -3.062,-2.799 -1.255,-0.935 -2.268,-2.438 -3.724,-3.072 -0.518,-0.226 -1.073,-0.345 -1.587,-0.58 -0.888,-0.406 -0.754,-1.063 -1.33,-1.673 -0.756,-0.801 -2.096,-0.853 -2.859,-1.588 -0.96,-0.924 0.317,-2.541 -0.708,-3.378 -1.235,-1.009 -1.471,-1.019 -2.015,-2.608 -0.464,-1.349 -2.756,-3.145 -3.831,-4.097"
-              id="path2204"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2047.038,799.272 c 1.494,1.597 2.011,-2.856 1.348,-3.759 -0.782,-1.063 -2.03,-3.071 -3.265,-3.606 -0.913,-0.396 -1.437,0.892 -2.403,0.308 0.153,1.342 1.342,2.333 1.917,3.489 0.357,0.717 1.186,0.965 1.604,1.628 0.379,0.603 0.448,1.326 0.799,1.94"
-              id="path2208"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2044.229,786.888 c -0.012,-0.72 -0.543,-1.907 -1.44,-1.224 0.542,0.334 1.022,0.742 1.44,1.224"
-              id="path2210"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2053.229,808.703 c 0.191,-0.638 -0.058,-1.267 -0.58,-1.66 -0.397,0.541 -0.689,1.375 -0.844,2.023 -0.207,0.868 1.711,0.901 1.424,-0.363"
-              id="path2212"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2051.928,804.603 c 0.377,1.068 0.941,2.047 1.511,3.021 0.245,-0.798 -0.386,-3.15 -1.511,-3.021"
-              id="path2214"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2053.729,790.56 c 0.107,-0.562 -0.133,-0.562 -0.721,0 0.24,0 0.48,0 0.721,0"
-              id="path2216"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2055.598,803.813 c -0.196,-0.952 -1.701,-0.8 -2.229,-0.291 0.668,0.445 1.469,0.499 2.229,0.291"
-              id="path2218"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2057.838,820.223 c 2.987,-0.421 -1.668,-6.247 -1.59,-7.63 -0.491,0.269 -1.323,0.138 -1.671,0.625 -0.5,0.701 -0.819,2.05 -0.287,2.802 1.009,1.422 1.594,3.783 3.548,4.203"
-              id="path2220"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2063.378,825.192 c -0.187,-0.771 -1.17,-3.485 -2.229,-2.762 -1.168,0.799 1.697,2.445 2.229,2.762"
-              id="path2222"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2064.389,828.653 c 0.227,0.854 1.222,0.459 1.526,-0.085 0.454,-0.81 -1.032,-2.033 -1.606,-2.365 -0.304,0.823 0.234,1.617 0.08,2.45"
-              id="path2224"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2067.768,833.473 c 0.031,-0.551 -0.48,-0.965 -0.383,-1.505 0.126,-0.699 0.541,-0.534 0.332,-1.229 -0.151,-0.507 -0.773,-2.076 -1.546,-1.784 -1.155,0.437 1.043,4.059 1.597,4.518"
-              id="path2226"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2078.139,787.824 c -0.299,-1.424 -2.284,-1.862 -2.796,-0.323 -0.284,0.857 0.26,2.078 1.206,2.267 1.142,0.228 1.179,-1.283 1.59,-1.944"
-              id="path2228"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2083.468,791.567 c -1.072,-0.454 -2.308,-0.39 -3.319,-1.008 -0.199,1.054 0.694,1.533 1.574,1.694 0.864,0.158 1.328,0.955 1.955,1.474 0,-0.726 0.067,-1.466 -0.21,-2.16"
-              id="path2232"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2083.608,794.231 c -0.44,-0.748 -1.086,-1.589 -2.011,-1.728 -0.241,1.315 -0.111,3.668 1.58,3.959 0.266,-0.724 0.135,-1.513 0.431,-2.231"
-              id="path2234"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2087.998,797.833 c -0.498,-0.946 -1.467,-1.593 -2.474,-1.886 -0.56,-0.164 -1.071,-0.968 -1.435,-1.038 -0.422,-0.081 -0.8,2.311 -0.287,2.653 1.181,0.792 2.89,-0.022 4.056,0.99 0.047,-0.239 0.093,-0.479 0.14,-0.719"
-              id="path2236"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2088.868,797.253 c -0.305,-1.235 -4.221,-3.62 -3.68,-1.439 1.501,-0.672 2.151,1.924 3.68,1.439"
-              id="path2238"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2089.878,799.563 c -0.105,-0.513 -0.419,-0.803 -0.939,-0.871 -0.363,0.542 -0.293,1.022 0.21,1.44 0.266,-0.159 0.509,-0.349 0.729,-0.569"
-              id="path2240"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2090.878,795.743 c -0.202,-0.29 -0.466,-0.505 -0.79,-0.647 -0.087,0.607 0.212,0.85 0.79,0.647"
-              id="path2242"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2091.458,798.263 c -0.176,-0.567 -0.616,-0.943 -1.229,-0.72 0.021,0.601 0.38,1.117 0.869,1.44 0.12,-0.24 0.24,-0.48 0.36,-0.72"
-              id="path2244"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2093.118,806.973 c -0.526,-0.36 -2.479,-0.153 -1.683,0.782 0.488,0.572 1.597,-0.229 1.683,-0.782"
-              id="path2248"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2094.768,796.603 c 0.206,-0.753 -0.057,-0.896 -0.789,-0.43 0.262,0.143 0.526,0.286 0.789,0.43"
-              id="path2252"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2096.348,794.808 c -0.521,-1.201 -1.51,0.163 -1.51,0.878 0,0.855 1.617,-0.058 1.51,-0.878"
-              id="path2256"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2096.568,797.763 c 0.395,-0.718 0.201,-0.958 -0.58,-0.72 0.194,0.24 0.387,0.48 0.58,0.72"
-              id="path2258"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2100.309,795.743 c -1.031,-2.517 -3.911,-0.152 -2.709,0.295 0.803,0.298 1.16,-0.474 1.418,0.855 0.259,1.336 1.319,-0.591 1.291,-1.15"
-              id="path2262"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2097.358,798.553 c -0.01,-0.278 -0.076,-0.541 -0.21,-0.79 0.071,0.263 0.14,0.527 0.21,0.79"
-              id="path2264"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2098.948,807.623 c -0.344,0.275 -1.023,0.414 -1.439,0.549 -0.798,0.258 0.546,1.838 0.64,2.261 0.053,-0.72 0.421,-0.946 1.01,-0.5 0.194,-0.859 0.82,-1.705 -0.211,-2.31"
-              id="path2266"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2098.518,803.233 c -0.412,0.453 -0.246,0.716 0.5,0.789 -0.167,-0.263 -0.333,-0.525 -0.5,-0.789"
-              id="path2268"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2099.229,804.603 c 0.104,0.514 -0.802,2.405 0.285,2.266 0.607,-0.078 0.975,-0.409 1.582,-0.102 0.622,0.315 1.147,0.819 1.883,0.426 -0.383,-0.607 -3.041,-2.662 -3.75,-2.59"
-              id="path2270"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2099.948,797.973 c -0.251,-0.013 -0.49,0.037 -0.72,0.15 0.241,-0.05 0.48,-0.1 0.72,-0.15"
-              id="path2272"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2099.518,802.803 c -0.324,0.703 0.724,1.268 1.16,1.511 -0.366,-0.52 -0.753,-1.023 -1.16,-1.511"
-              id="path2274"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2105.498,819.503 c 0.844,1.158 -1.296,1.682 -1.73,2.45 0.884,0.716 1.558,1.301 2.745,0.865 0.8,-0.293 2.125,0.39 2.353,1.21 0.203,0.726 0.76,1.597 0.878,2.236 0.274,1.471 -0.593,2.088 1.081,3.005 0.719,0.394 3.973,2.894 4.754,2.184 0.776,-0.707 -0.414,-1.035 -0.649,-1.619 -0.372,-0.924 0.763,-2.595 1.659,-2.771 -1.139,-0.688 -2.702,-0.354 -3.75,-1.221 -0.965,-0.799 -1.307,-2.777 -1.201,-3.955 0.062,-0.7 -0.411,-2.962 -0.879,-3.535 -1.32,-1.625 -3.029,0.243 -1.81,1.73 -0.514,0.07 -3.349,-3.846 -3.451,-0.579"
-              id="path2282"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2116.518,829.733 c -0.217,0.07 -0.433,0.139 -0.649,0.209 0.148,0.597 0.599,0.856 1.149,0.511 -0.139,-0.259 -0.306,-0.499 -0.5,-0.72"
-              id="path2286"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2121.559,829.083 c -0.335,0.121 -0.625,0.311 -0.87,0.57 0.563,0.269 0.891,0.053 0.87,-0.57"
-              id="path2290"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2123.348,826.772 c -0.95,0.967 -1.174,2.638 -1.07,3.914 0.11,1.362 2.126,2.079 1.791,0.627 0.593,0.039 1.479,1.684 2.066,1.105 0.805,-0.794 1.628,-2.402 1.496,-3.569 -0.193,-1.707 -2.897,-2.355 -4.283,-2.077"
-              id="path2292"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2125.658,771.336 c 0.62,0.102 1.183,0.455 1.58,0.937 -0.584,-0.368 -0.97,-0.062 -1.01,0.575 1.788,0.257 2.295,-1.537 1.6,-2.891 -1.235,-2.403 -2.762,-0.496 -2.17,1.379"
-              id="path2294"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2130.838,819.503 c 0.302,0.028 0.592,-0.018 0.87,-0.14 -0.303,-0.028 -0.593,0.019 -0.87,0.14"
-              id="path2298"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2133.068,818.213 c -0.405,-0.32 -0.812,-0.32 -1.221,0 0.168,0.284 0.385,0.523 0.65,0.72 0.22,-0.218 0.409,-0.458 0.571,-0.72"
-              id="path2302"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2139.188,813.603 c -1.129,-1.601 -3.427,0.908 -2.011,2.301 0.678,-0.677 2.465,-0.985 2.011,-2.301"
-              id="path2306"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2188.588,833.043 c -2.61,-1.094 -1.773,5.536 -1.229,6.41 0.688,-1.134 1.427,-0.789 1.328,-2.342 -0.078,-1.218 -0.575,-2.905 -0.099,-4.068"
-              id="path2324"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2189.738,834.123 c -0.814,0.149 -0.746,1.017 -0.221,1.44 0.117,-0.473 0.191,-0.953 0.221,-1.44"
-              id="path2326"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2197.438,775.8 c -0.156,0.657 0.084,0.729 0.72,0.216 -0.24,-0.072 -0.479,-0.144 -0.72,-0.216"
-              id="path2332"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2198.309,774.936 c -0.313,0.695 0.353,1.305 0.93,1.513 0.047,-0.649 0.063,-1.743 -0.93,-1.513"
-              id="path2334"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2198.519,769.176 c -0.016,0.621 0.382,1.037 1.01,0.72 -0.074,-0.578 -0.434,-0.858 -1.01,-0.72"
-              id="path2336"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2198.809,807.843 c -0.226,0.208 -0.322,0.472 -0.29,0.79 0.096,-0.263 0.193,-0.527 0.29,-0.79"
-              id="path2338"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2200.038,773.712 c 0.026,0.517 0.29,0.853 0.79,1.008 0.017,-0.526 -0.302,-0.869 -0.79,-1.008"
-              id="path2340"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2206.298,784.584 c 0.053,0.518 0.293,0.925 0.721,1.224 -0.218,-0.421 -0.458,-0.829 -0.721,-1.224"
-              id="path2344"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2132.639,797.113 c 0.934,-0.626 0.299,-1.282 0.499,-2.125 0.261,-1.095 1.85,-1.166 2.021,-2.413 -0.331,0.544 -0.737,1.023 -1.22,1.439 -0.878,-1.292 0.21,-2.565 1.038,-3.522 0.54,-0.623 0.701,-1.208 0.982,-1.959 0.364,-0.973 1.61,-1.199 2.299,-1.862 -0.132,0.89 -1.308,1.981 0.034,2.517 0.191,0.077 0.672,2.179 1.049,2.59 1.103,1.202 3.083,2.143 3.791,3.677 0.274,0.593 1.497,1.107 2.085,1.412 1.102,0.57 2.655,-0.557 3.606,-0.98 1.89,-0.843 3.688,0.271 5.701,0.075 0.805,-0.079 1.536,-0.697 2.246,-1.076 1.062,-0.567 0.761,-1.953 1.565,-2.828 0.838,-0.912 5.094,-1.836 5.752,-0.441 0.627,1.327 4.512,2.593 5.754,1.827 1.073,-0.662 1.304,-1.717 2.76,-1.216 0.716,0.247 1.285,-0.34 1.9,-0.585 0.509,-0.205 1.104,0.456 1.626,0.43 -0.281,-1.386 2.66,-4.404 1.852,-5.299 -1.394,-1.544 1.732,-3.514 2.898,-3.917 -0.684,-0.845 -1.969,-3.955 0.162,-4.211 2.677,-0.323 0.672,-1.98 2.204,-3.447 0.591,-0.565 -0.171,-2.877 -0.08,-3.85 0.084,-0.9 0.316,-2.04 0.721,-2.851 0.35,-0.698 1.066,-1.031 1.243,-1.842 1.416,0.527 1.769,0.808 3.278,0.159 0.514,-0.221 1.778,1.684 1.902,-0.304 0.978,0.944 2.825,0.707 4.045,0.595 1.563,-0.143 2.515,1.556 4.024,1.422 -0.454,0.321 -0.957,0.513 -1.51,0.575 0.272,0.237 2.981,2.492 1.15,1.585 0.372,0.262 1.333,0.437 1.521,0.873 0.188,0.437 0.176,0.872 0.709,1.07 -0.863,1.107 -2.271,0.415 -2.66,-0.72 0.178,0.489 0.248,0.993 0.21,1.512 -0.777,-0.482 -2.138,-0.433 -2.95,-0.072 1.029,0.398 1.561,1.454 2.671,1.729 -0.283,0.566 -0.413,1.815 -1.37,1.368 0.705,0.917 1.756,1.3 2.81,1.655 -0.876,1 0.819,1.098 -0.58,1.296 0.483,0.019 0.94,0.139 1.37,0.36 -1.428,0.39 0.364,1.203 0.908,1.871 0.621,0.765 1.171,1.6 1.682,2.442 0.643,1.059 -0.711,3.142 -1.95,3.103 0.808,0.284 0.412,1.089 1.042,1.612 0.907,0.753 1.813,1.53 2.758,2.235 0.762,0.568 1.211,1.54 2.101,1.86 1.192,0.429 2.03,1.643 2.82,2.573 0.333,0.392 1.2,0.084 0.595,1.031 -0.748,1.17 -1.968,0.589 -3,1.112 -1.712,0.866 -3.735,-1.519 -5.085,-2.066 0.623,0.912 1.114,2.063 -0.344,2.146 -1.324,0.074 -1.743,2.093 -2.08,3.085 -0.901,2.65 -1.334,5.686 -1.176,8.449 0.381,-0.438 0.86,-0.678 1.439,-0.72 -0.185,0.594 -1.138,1.304 -0.43,1.58 -0.682,0.422 -0.682,1.017 0,1.439 -0.936,0.264 -1.976,1.382 -2.23,0.07 -0.289,1.857 -3.743,5.774 -4.25,2.02 -0.724,1.313 0.504,2.341 -0.977,3.37 -0.683,0.474 -0.346,1.213 -0.823,1.811 -0.555,0.694 -1.581,0.671 -2.09,1.439 2.264,-0.859 1.513,2.29 0.43,3.101 0.559,0.549 1.704,-0.212 2.174,0.356 0.456,0.553 -0.463,2.69 -0.804,3.173 -0.396,-0.466 -0.875,-0.585 -1.44,-0.359 1.417,1.226 -0.438,5.153 -1.439,3.16 -0.27,0.875 0.463,1.628 0.939,2.239 -1.98,0.363 -1.371,3.374 -3.038,4.048 -1.828,0.74 -3.673,1.97 -5.531,2.792 -0.489,0.216 -4.347,2.829 -4.184,1.356 0.099,-0.894 -0.01,-1.875 0.034,-2.787 0.059,-1.3 -0.781,-2.377 -0.461,-3.689 -1.462,1.433 -2.462,0.893 -1.439,-1.08 -0.663,1.122 -2.132,1.844 -1.801,-0.141 -0.154,1.053 -2.934,3.5 -3.744,1.877 -0.28,-0.561 0.495,-0.992 -0.042,-1.582 -0.987,-1.083 -2.039,0.879 -2.333,-0.585 0.325,1.666 -1.816,-1.299 -2.021,-2.02 0.013,1.508 -1.611,3.187 -2.892,3.842 -1.333,0.682 -2.5,-1.237 -3.793,-0.304 -0.469,0.338 -2.285,2.947 -2.637,1.253 -0.428,-2.062 0.27,-3.904 -0.828,-5.941 -0.863,2.233 -1.163,1.193 -2.814,1.147 -0.695,-0.019 -1.264,0.582 -1.802,0.922 -1.023,0.648 -1.693,0.069 -2.725,-0.119 0.334,-0.364 0.694,-0.701 1.08,-1.01 -0.835,0.126 -1.108,1.208 -2.02,1.08 0.382,-2.191 -1.826,-0.977 -2.73,-0.29 -1.05,-1.31 -0.211,-3.482 -0.673,-4.993 -0.313,-1.027 -0.122,-2.113 -0.311,-3.157 -0.159,-0.881 -1.537,-1.589 -1.446,-2.36 0.021,-0.181 0.779,-1.114 0.87,-1.383 0.322,-0.962 0.498,-1.796 -0.093,-2.683 -0.404,-0.606 -0.489,-1.259 -0.833,-1.841 -0.419,-0.708 -1.315,-0.97 -1.265,-1.943 -0.206,0.25 -0.423,0.489 -0.649,0.72 -0.763,-1.473 -1.574,-0.118 -2.616,-0.748 -0.59,-0.356 -0.216,-1.014 0.044,-1.417 0.262,-0.406 1.229,0.062 1.563,0.075 -0.563,-0.314 -0.898,-1.078 -1.521,-1.234 -0.463,-0.117 -1.061,-0.03 -1.363,-0.495 -0.425,-0.652 -0.675,-3.309 0.653,-2.511 -0.345,-0.928 -0.918,-2.051 0.22,-2.159 -0.961,-1.269 -2.236,-2.372 -2.577,-3.985 -0.142,-0.674 -0.767,-2.861 -0.16,-3.433"
-              id="path2554"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2195.788,774.144 c 0.023,-0.168 0.047,-0.336 0.07,-0.504 0.769,0.11 1.483,0.427 2.16,0.792 -0.758,0.138 -1.517,-0.04 -2.23,-0.288"
-              id="path2556"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2199.098,768.6 c 0.875,-0.116 2.018,-0.281 1.801,1.009 -0.836,0.174 -1.252,-0.574 -1.801,-1.009"
-              id="path2558"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2324.378,856.803 c 0.05,0.594 0.289,1.098 0.72,1.511 -0.195,-0.343 0.149,-2.436 0.301,-2.909 0.386,-1.207 -0.885,-0.581 -0.521,-1.632 0.98,0.492 1.611,-1.117 2.024,-1.7 0.708,-1.001 1.637,1.145 1.624,1.688 -0.034,1.421 -0.126,2.947 -0.207,4.368 -0.129,2.229 -1.793,1.207 -2.932,0.405 1.469,1.748 -3.815,-1.891 -1.009,-1.731"
-              id="path2564"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2326.108,859.033 c 0.3,0.197 1.239,1.418 0.621,1.71 -1.006,0.476 -1.685,-1.459 -0.621,-1.71"
-              id="path2566"
-              data-inkscape-connector-curvature="0"
-            />
-          </g>
-          <g
-            transform="matrix(6.4876122,0,0,6.4876122,-12925.057,-3966.0743)"
-            style={{
-              fill: clickedCountryId === "Thailand" ? "#FF0000" : "#005500",
-              fillOpacity: 0.98823529,
-              stroke: "#ffffff",
-              strokeWidth: 0.5257113,
-              strokeMiterlimit: 3.98288536,
-              strokeDasharray: "none",
-              strokeOpacity: 1,
-            }}
-            id="Thailand"
-            className="country"
-          >
-            <path
-              d="m 2049.559,733.824 c -0.371,1.208 -0.162,2.515 0.279,3.672 1.072,-0.862 0.953,-2.997 -0.279,-3.672"
-              id="path3902"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2051.348,734.472 c 0.217,0.503 0.477,0.99 0.65,1.512 0.041,-0.615 -0.176,-1.119 -0.65,-1.512"
-              id="path3904"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2055.958,740.088 c -0.083,-0.578 -0.323,-1.082 -0.72,-1.512 0.062,0.589 0.302,1.093 0.72,1.512"
-              id="path3906"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2060.568,746.64 c -0.509,0.51 -0.562,1.265 0.29,1.44 -0.091,-0.481 -0.188,-0.961 -0.29,-1.44"
-              id="path3908"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2061.798,721.8 c -0.497,0.453 -0.474,0.885 0.07,1.296 0.484,-0.448 0.46,-0.88 -0.07,-1.296"
-              id="path3910"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2061.648,719.928 c -0.222,0.513 0.195,0.726 0.649,0.72 -0.195,-0.259 -0.412,-0.499 -0.649,-0.72"
-              id="path3912"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2077.639,700.776 c 0.236,0.023 0.473,0.048 0.709,0.071 0.01,-0.583 -0.354,-1.047 -0.859,-1.296 0,0.415 0.055,0.823 0.15,1.225"
-              id="path3914"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2079.868,702.936 c -0.229,0.561 -0.108,1.041 0.36,1.44 0.142,-0.545 0.022,-1.025 -0.36,-1.44"
-              id="path3916"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2034.288,643.824 c 0.445,0.857 2.298,0.823 2.909,0.195 0.737,-0.756 -0.72,-2.807 -0.529,-3.796 0.106,-0.551 0.544,-0.931 0.57,-1.512 0.041,-0.9 -0.688,-1.483 -0.444,-2.442 0.154,-0.611 1.266,-2.68 1.682,-2.604 1.22,0.226 2.414,0.754 3.667,0.48 0.595,-0.129 1.165,-0.808 1.756,-0.795 0.601,0.013 0.837,0.599 1.345,-0.178 0.499,-0.765 -0.431,-2.221 0.747,-2.524 0.846,-0.217 4.067,0.439 2.338,-1.801 1.234,-0.062 4.552,-2.249 4.54,0.433 0.677,-0.188 0.892,-1.157 1.684,-0.92 0.353,0.105 1.548,1.696 1.832,2.021 0.696,0.792 -0.282,1.98 -0.487,2.773 -0.288,1.118 0.488,1.991 1.297,2.654 1.135,0.926 2.797,-0.936 4.095,-0.552 0.788,0.232 1.517,0.238 1.083,1.288 -0.417,1.011 0.992,2.461 1.158,3.398 0.141,0.793 -0.38,1.579 -0.269,2.419 0.147,1.117 -0.854,1.695 -0.807,2.634 0.023,0.452 0.834,0.716 0.921,1.301 0.145,0.974 0.197,1.834 -0.164,2.758 -0.879,2.252 -1.671,3.542 0.957,4.489 1.088,-1.697 3.853,-2.961 3.96,-5.184 1.221,0.731 1.315,-1.256 2.25,-1.274 0.646,-0.012 1.515,1.247 2.104,1.54 0.461,0.23 1.526,0.311 1.808,0.784 0.893,1.494 1.934,-0.698 3.059,-0.474 0.608,0.121 0.826,-2.312 1.801,-2.448 -1.685,-1.856 3.48,-0.653 4.038,-0.498 3.138,0.872 4.25,5.244 7.157,6.768 2.402,1.26 1.226,5.698 1.674,7.917 0.272,1.347 1.469,2.043 2.229,3.102 1.222,1.699 1.786,1.692 3.691,2.439 -1.625,1.109 0.798,1.867 1.532,2.453 0.856,0.683 -0.819,3.523 0.338,3.307 -0.313,0.713 -1.03,1.405 -0.536,2.214 0.28,0.458 0.845,0.663 0.785,1.287 -0.165,1.717 0.509,3.413 -1.204,4.477 -1.144,0.71 -1.763,2.196 -2.649,0.597 -0.914,-1.646 -3.513,-0.017 -4.862,-0.422 -0.936,-0.28 -1.834,0.058 -2.756,0.187 -0.989,0.138 -1.447,-0.471 -2.318,-0.542 -1.201,-0.099 -2.689,0.449 -3.791,0.882 -1.74,0.684 -1.876,2.747 -2.759,4.138 -0.409,0.646 -1.479,1.178 -1.388,1.993 0.098,0.862 -1.022,0.545 -1.563,0.526 0.557,1.467 0.791,3.053 1.33,4.544 0.45,1.246 0.383,2.908 1.584,3.75 1.349,0.945 1.021,1.998 1.139,3.418 0.146,1.74 2.044,3.159 1.777,4.921 -0.018,-0.881 -0.752,-1.361 -1.14,-2.065 -0.585,-1.063 -0.507,-2.399 -1.95,-2.832 0.033,0.432 0.057,0.863 0.07,1.296 -0.398,-0.955 -3.621,-1.404 -1.73,-2.664 -0.521,0.311 -0.886,-0.021 -1.08,-0.504 0,1.551 -1.835,-0.848 -2.159,-1.584 0,1.364 -1.03,-1.4 -2.314,-0.813 -0.626,0.286 -1.259,0.817 -2.006,0.668 -0.915,-0.183 -1.972,-0.686 -2.919,-0.452 -1.852,0.458 -0.813,-1.462 -0.937,-2.603 -0.072,-0.668 -0.094,-1.345 -0.187,-2.011 -0.127,-0.921 0.55,-1.639 0.372,-2.494 -0.784,0.628 -2.879,0.048 -3.17,-0.936 0.341,1.974 -5.648,0.23 -3.935,3.297 0.854,1.529 0.311,2.501 0.01,4.099 -0.322,1.696 0.939,3.281 0.619,4.988 -0.127,0.677 -0.63,1.188 -0.969,1.758 -0.417,0.698 -0.153,1.543 -0.632,2.231 -0.901,1.298 -1.042,3.293 -1.371,4.807 -0.401,1.846 -0.671,3.178 -1.442,4.873 -0.275,0.605 0.305,2.268 -0.847,2.099 1.303,1.574 -0.079,4.487 0.745,6.39 0.393,0.905 0.975,1.508 0.521,2.428 -0.513,1.042 1.382,1.274 1.928,0.985 1.923,-1.018 3.105,-1.039 3.396,1.574 0.096,0.862 1.313,7.038 2.561,5.04 0.762,1.756 0.888,3.708 1.468,5.521 0.52,1.621 1.324,3.332 1.342,5.063 -1.195,-0.218 -1.43,-5.458 -2.722,-4.553 -0.929,0.651 0.456,2.702 0.82,3.226 0.435,0.623 0.948,1.103 1.501,1.61 0.461,0.423 1.136,0.262 1.586,0.684 1.043,0.98 1.244,2.003 2.967,2.053 1.266,0.036 2.678,-0.466 3.791,0.385 1.225,0.936 1.484,2.647 2.67,3.639 1.254,1.049 2.063,1.554 1.436,3.216 -0.288,0.762 -0.791,2.31 -1.739,2.511 -0.729,0.155 -0.873,-1.037 -1.43,-1.257 -1.46,-0.578 -2.271,1.999 -3.205,1.994 -1.803,-0.009 -1.355,-1.782 -0.725,-2.771 0.686,-1.074 -0.542,-2.057 -1.329,-2.251 -1.117,-0.275 -0.671,-1.867 -2.1,-2.059 -1.315,-0.176 -2.079,-0.231 -2.922,-1.333 -1.181,-1.545 -0.761,1.124 -1.27,1.819 -0.11,-2.04 -5.207,-3.954 -3.17,-6.12 -1.716,1.03 -1.028,-1.643 -1.87,-2.159 0.088,1.049 -0.797,0.826 -1.179,0.187 -0.492,-0.825 -0.382,-2.077 -1.096,-2.764 -0.473,-0.455 -1.418,0.138 -1.511,-0.609 -0.084,-0.679 0.056,-0.929 -0.345,-1.496 -0.604,-0.854 -0.532,-0.789 -1.428,-0.96 -0.561,-0.106 -0.455,-1.027 -0.632,-1.414 -0.286,-0.627 -0.611,-1.593 -1.511,-1.368 -0.96,0.24 -0.563,1.117 -0.869,1.801 -2.407,-0.496 -1.926,-4.098 -1.651,-5.836 0.554,-3.5 1.295,-7.056 2.036,-10.523 0.287,-1.346 1.396,-3.134 1.38,-4.458 -0.02,-1.556 0.586,-1.645 1.181,-3.031 0.749,-1.745 5.33,-6.426 3.78,-8.347 -0.663,-0.818 -1.216,-3.044 -1.454,-4.047 -0.362,-1.528 -0.925,-2.047 -1.68,-3.339 -0.435,-0.744 -1.413,-2.831 -1.267,-3.643 0.354,-1.962 0.27,-3.39 -0.53,-5.277 -1.303,-3.071 -4.336,-4.927 -6.303,-7.553 -0.861,-1.151 -1.702,-2.284 -1.886,-3.771 -0.104,-0.836 0.142,-2.758 1.353,-1.664 -0.261,-1.018 0.943,-0.8 1.19,-1.514 0.352,-1.013 -0.289,-3.383 -0.447,-4.471 -0.262,-1.798 1.146,-0.878 1.781,-1.885 0.407,-0.646 0.427,-2.043 -0.214,-2.571 -0.181,0.567 -0.471,1.071 -0.87,1.513 -1.214,-1.503 -1.563,-3.26 -2.08,-5.04 -0.328,-1.133 -2.003,-2.254 -2.794,-3.136 -1.219,-1.358 -2.827,-2.797 -3.854,-4.278 -0.568,-0.82 -0.304,-1.593 -0.574,-2.479 -0.245,-0.799 -1.093,-1.433 -0.917,-2.347 -1.491,0.437 -2.145,-1.528 -2.449,-2.594"
-              id="path3918"
-              data-inkscape-connector-curvature="0"
-            />
-          </g>
-          <g
-            transform="matrix(6.4876122,0,0,6.4876122,-12925.057,-3966.0743)"
-            style={{
-              fill: clickedCountryId === "Myanmar" ? "#FF0000" : "#005500",
-              fillOpacity: 1,
-              stroke: "#ffffff",
-              strokeWidth: 0.5257113,
-              strokeMiterlimit: 3.98288536,
-              strokeDasharray: "none",
-              strokeOpacity: 1,
-            }}
-            id="Myanmar"
-            className="country"
-          >
-            <path
-              d="m 1997.858,627.768 c -0.05,0.371 0.116,0.636 0.5,0.792 -0.167,-0.264 -0.333,-0.529 -0.5,-0.792"
-              id="path4568"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 1999.438,631.584 c 0.333,0.479 0.623,0.983 0.87,1.512 -0.21,-0.819 -0.559,-1.62 -1.16,-2.231 0.097,0.239 0.194,0.479 0.29,0.719"
-              id="path4570"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2003.118,632.016 c -0.253,0.578 0.492,1.23 0.86,1.513 -0.023,-0.616 -0.379,-1.153 -0.86,-1.513"
-              id="path4572"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2005.918,641.231 c -0.394,0.021 -0.753,0.141 -1.08,0.36 2.338,2.691 2.365,-0.41 1.08,-0.36"
-              id="path4574"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2007.868,637.056 c -0.421,-0.847 -1.465,-0.877 -2.021,-1.584 -1.045,1.065 1.05,3.05 2.021,1.584"
-              id="path4576"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2016.218,664.056 c -0.484,0.531 -2.258,2.198 -1.229,2.952 0.738,-0.779 1.812,-1.762 1.229,-2.952"
-              id="path4578"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2017.878,666.864 c -0.194,0.342 -0.348,0.237 0,0.72 0,-0.24 0,-0.48 0,-0.72"
-              id="path4580"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2037.678,661.824 c -0.389,0.172 -0.513,1.728 -0.015,1.728 0.975,0 0.944,-1.607 0.015,-1.728"
-              id="path4582"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2044.729,698.256 c 0.207,-0.224 0.327,-0.487 0.359,-0.792 -0.855,-0.141 -0.975,0.123 -0.359,0.792"
-              id="path4584"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2045.668,703.728 c 0.023,-0.543 -0.146,-1.022 -0.51,-1.439 0.205,0.466 0.375,0.946 0.51,1.439"
-              id="path4586"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2045.448,699.336 c 0.144,0 0.286,0 0.43,0 -0.182,-0.292 -0.398,-0.557 -0.649,-0.792 0.073,0.264 0.146,0.528 0.219,0.792"
-              id="path4588"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2046.458,692.352 c 0.456,-0.575 0.202,-1.168 -0.43,-1.439 0.139,0.481 0.282,0.96 0.43,1.439"
-              id="path4590"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2048.038,711.864 c -0.094,-0.843 -0.505,-1.613 -1.22,-2.088 -0.187,0.214 -0.329,0.454 -0.43,0.72 0.986,-0.477 1.352,0.714 1.65,1.368"
-              id="path4592"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2046.678,705.528 c 1.211,-0.083 1.095,-2.277 0.57,-2.952 -0.247,0.971 -0.437,1.96 -0.57,2.952"
-              id="path4594"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2047.178,697.608 c 0.368,0.525 1.997,-2.688 -0.289,-2.232 0.14,0.737 0.442,1.466 0.289,2.232"
-              id="path4596"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2047.538,718.128 c 0.047,0.24 0.093,0.479 0.14,0.72 0.449,-0.286 0.666,-0.694 0.65,-1.224 -0.28,0.142 -0.543,0.31 -0.79,0.504"
-              id="path4598"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2048.479,704.16 c 0.168,0.416 0.479,0.632 0.93,0.647 0.359,-2.339 -2.246,-2.784 -0.93,-0.647"
-              id="path4600"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2048.398,700.776 c 0.197,0.334 0.391,0.67 0.58,1.008 0.283,-0.168 0.522,-0.385 0.72,-0.648 -0.58,-0.226 -1.061,-0.586 -1.44,-1.08 0.047,0.24 0.094,0.48 0.14,0.72"
-              id="path4602"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2048.328,697.176 c 0.145,-0.386 0.389,-0.243 0.07,-0.792 -0.023,0.263 -0.046,0.528 -0.07,0.792"
-              id="path4604"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 1992.528,621.288 c 0.071,-0.554 -0.042,-1.955 0.777,-2.047 1.015,-0.114 1.668,1.417 2.683,1.111 -0.536,-1.286 -0.909,-2.644 -1.045,-4.033 0.031,-0.572 -0.029,-1.133 -0.18,-1.685 -0.215,-0.535 0.635,-1.083 0.504,-1.698 0.733,0.262 1.138,1.815 2.046,1.53 0.65,-0.204 0.352,-1.48 0.475,-1.963 1.51,0.881 0.864,-1.338 0.587,-1.999 -0.779,-1.86 -0.622,-3.305 -0.877,-5.273 1.486,1.103 1.543,-1.167 1.516,-2.103 -0.068,-2.307 -0.502,-4.292 -1.296,-6.465 0.346,-0.025 0.651,0.075 0.917,0.299 0.683,0.458 0.64,0.123 1.457,-0.022 0.814,-0.146 3.252,0.876 3.887,1.38 -0.028,-3.304 4.521,-9.193 1.958,-11.766 -1.36,-1.365 1.522,-3.268 1.832,-4.632 0.36,-1.589 0.84,-2.508 0.048,-3.915 -0.651,-1.158 -0.851,-3.383 0.412,-4.24 0.391,-0.267 1.579,0.086 1.65,-0.606 0.065,-0.646 1.356,-1.075 1.854,-1.549 0.629,-0.6 0.972,-1.21 1.393,-1.961 0.714,-1.277 2.569,-1.091 3.813,-1.283 0.598,-0.092 1.328,-0.511 1.856,-0.006 0.713,0.681 1.447,2.271 2.753,1.733 -0.203,-1.106 -1.289,-1.641 -1.958,-2.429 -1.224,-1.441 -0.405,-2.551 0.691,-3.648 1.576,-1.574 0.837,-1.257 0.837,-2.958 0,-0.603 0.756,-0.835 0.699,-1.393 -0.07,-0.693 0.208,-1.342 0.23,-2.027 1.346,1.496 4.048,1.738 5.138,3.649 0.396,0.693 0.272,1.548 0.623,2.255 0.476,0.958 1.183,1.778 1.659,2.736 0.56,-0.199 0.864,-0.649 0.57,-1.225 2.986,1.388 3.058,4.47 3.872,7.424 0.401,1.457 0.678,2.685 0.444,4.188 -0.096,0.617 1.024,2.688 -0.566,2.068 0.244,0.624 2.298,3.043 0.257,2.432 -0.365,-0.109 -1.015,1.448 -0.953,1.743 0.198,0.94 -1.041,0.325 -1.533,0.29 -0.15,0.529 -0.264,4.479 -1.95,2.952 0.071,1.332 -0.629,2.328 0.149,3.6 -1.341,0.105 -1.335,1.074 -1.279,2.126 0.057,1.09 0.947,1.103 1.326,1.947 0.165,0.368 0.766,1.388 0.599,1.802 -0.29,0.72 -1.393,1.747 -0.475,2.472 0.429,0.338 2.888,-1.609 3.307,-1.844 1.814,-1.02 3.984,0.619 5.732,-0.67 0.278,1.125 -1.188,1.568 -0.815,2.686 0.211,0.634 0.637,0.95 1.068,1.418 0.546,0.591 -0.132,1.251 0.249,1.578 1.026,0.882 0.636,2.517 1.966,3.194 1.813,0.923 3.281,-0.244 4.013,1.924 -1.86,-0.075 -0.844,2.162 -0.714,3.172 0.137,1.061 -2.128,3.149 -0.207,3.765 1.308,0.419 4.075,-0.451 5.171,0.551 1.198,1.097 -0.477,3.208 1.94,3.24 -0.329,0.502 -0.069,1.426 0.275,1.853 0.559,0.693 1.086,-0.815 1.865,-0.313 1.607,1.036 3.7,-1.498 4.919,-2.403 0.087,1.255 1.046,1.957 -0.119,3.002 -0.537,0.482 -1.773,0.758 -2.088,1.313 -0.247,0.436 -1.788,4.192 -0.313,3.317 -0.114,1.069 -1.821,0.53 -2.449,1.008 -1.096,0.834 -0.724,2.505 -1.011,3.672 -1.458,-1.434 -2.612,0.2 -4.39,0.288 0.469,0.606 0.463,1.611 -0.323,1.964 -0.778,0.35 -1.253,-0.466 -1.976,-0.407 -0.9,0.075 -0.688,2.034 -0.852,2.624 -0.184,0.661 -0.568,0.582 -1.043,0.261 -0.468,-0.316 -1.667,1.063 -2.356,0.995 -0.346,-0.034 -3.74,-0.532 -3.74,-0.623 0,1.169 -0.937,1.952 -1.245,3.117 -0.151,0.57 0.28,1.256 0.41,1.79 0.175,0.716 -0.539,1.475 -0.633,2.19 -0.122,0.933 1.307,2.015 0.662,2.982 -0.581,0.874 -2.401,1.129 -2.944,0.084 0.304,1.065 0.958,3.031 2.45,2.592 -0.13,0.679 0.324,1.223 0.643,1.759 0.474,0.8 0.212,1.595 0.508,2.417 0.511,1.423 2.549,3.119 3.565,4.237 0.622,0.684 1.26,1.355 1.765,2.135 0.462,0.714 1.33,0.558 1.659,1.692 0.518,1.78 0.866,3.537 2.08,5.04 0.399,-0.441 0.689,-0.945 0.87,-1.513 0.909,0.749 0.584,2.266 -0.22,2.952 -1.068,0.912 -1.67,-0.035 -1.289,1.848 0.216,1.069 0.448,2.769 0.334,3.845 -0.09,0.853 -1.394,0.788 -1.136,1.796 -0.697,-0.629 -1.252,-0.138 -1.406,0.629 -0.22,1.099 0.375,2.489 0.897,3.426 1.399,2.511 3.949,4.292 5.873,6.339 1.811,1.928 2.142,4.541 2.382,7.03 0.11,1.15 -0.596,1.013 -0.047,2.309 0.52,1.229 0.594,1.932 1.578,2.9 0.925,0.91 0.766,1.932 1.146,3.071 0.388,1.156 1.108,2.132 1.446,3.318 0.481,1.692 -1.115,3.966 -1.883,5.508 -0.803,1.61 -1.826,3.471 -3.132,4.556 -0.907,0.754 0.09,1.603 -0.352,2.487 -0.652,1.299 -0.768,2.78 -1.533,4.075 -0.861,-1.354 -0.422,-3.344 -0.78,-4.863 -0.146,-0.621 -0.479,-1.156 0.136,-1.538 0.752,-0.465 0.178,-0.744 0.539,-1.322 0.777,-1.245 0.947,-2.736 0.65,-4.197 -0.141,-0.692 -0.566,-2.989 0.685,-2.985 -0.199,-0.333 -0.296,-0.693 -0.29,-1.08 -0.349,0.483 -0.858,0.859 -1.439,1.008 -0.179,-1.028 1.237,-1.588 -0.221,-1.439 1.667,-0.794 -0.775,-2.05 0.58,-2.809 -0.597,0.158 -1.1,-0.151 -1.3,-0.72 0.395,-0.059 0.755,-0.203 1.08,-0.433 -0.652,-0.352 -1.103,-1.133 -0.069,-1.367 -0.889,-0.295 -0.352,-1.159 -0.486,-1.959 -0.224,-1.33 -0.728,-2.728 -1.24,-3.973 -0.539,-1.309 -1.333,-2.562 -1.968,-3.831 -0.604,-1.209 -0.661,-2.669 -1.486,-3.773 0.382,1.029 0.856,3.382 0.07,4.319 0.277,-0.853 -0.456,-1.616 -0.528,-2.437 -0.069,-0.774 -0.092,-2.042 -0.271,-2.747 -0.17,-0.669 -2.918,-5.397 -1.221,-4.464 -0.07,-0.286 -0.19,-0.55 -0.359,-0.792 -0.24,0.144 -0.48,0.288 -0.721,0.432 -1.239,-3.095 -0.962,-6.559 -2.096,-9.643 -0.22,-0.597 -0.754,-0.921 -1.017,-1.473 -0.234,-0.492 0.034,-1.071 0.01,-1.587 -0.05,-0.944 -0.72,-2.788 0.517,-3.28 -0.796,0.127 -2.195,0.678 -2.789,-0.099 -1.082,-1.417 -1.939,-2.975 -2.041,-4.797 -0.71,0.474 -0.572,-0.565 -1.177,-1.121 -0.522,-0.479 -1.195,-0.894 -1.413,-1.615 -0.146,0.183 0.113,1.736 0.107,2.09 -0.018,1.219 0.722,2.466 0.072,3.741 -0.469,0.924 -1.067,2.206 -2.26,2.304 -1.225,0.102 -2.63,-1.555 -1.729,-2.664 -1.423,0.357 -0.192,1.404 0.188,2.279 0.526,1.211 -0.751,2.596 -1.849,1.465 1.174,0.826 -1.157,2.293 -2.021,1.439 0.293,1.104 -0.583,4.851 -2.395,4.164 -0.935,-0.354 -0.066,-2.843 -0.265,-3.588 -0.045,0.068 -0.983,3.496 -0.5,1.224 -0.296,0.615 -0.416,1.264 -0.36,1.944 -0.899,-0.443 -1.147,-2.403 -0.439,-2.952 -1.064,-0.695 -0.644,4.324 -2.076,3.074 -0.682,-0.595 -0.088,-3.847 0.496,-4.154 -1.05,0.413 -1.338,2.563 -1.87,3.528 -0.313,-0.186 -0.576,-0.426 -0.79,-0.72 0.881,-0.369 1,-1.279 1.22,-2.088 -0.413,0.707 -0.581,1.765 -1.58,1.871 1.084,-1.461 0.59,-3.298 0.79,-4.968 -0.144,0.014 -0.286,0.014 -0.43,0 0.56,1.202 -1.072,3.892 -2.191,4.342 -1.618,0.651 -0.307,-6.197 -0.099,-6.956 0.209,-0.762 -0.022,-3.011 0.4,-3.411 1.072,-1.014 -0.098,-3.866 -0.603,-4.984 -0.649,-1.438 -0.431,-3.623 -1.497,-4.802 -0.452,-0.5 -0.714,-1.151 -1.006,-1.748 -0.438,-0.895 -1.623,-1.423 -0.405,-1.665 -0.85,-0.587 -1.87,-0.785 -1.87,-2.027 0,-0.925 0.054,-2.412 -0.439,-3.229 0.118,0.673 0.416,2.223 -0.79,1.367 0.556,0.682 1.005,2.039 0.58,2.881 -1.742,-0.709 -2.807,-2.511 -3.831,-3.989 -0.281,-0.404 0.605,-0.815 0.904,-0.266 0.224,0.413 0.814,0.557 1.229,0.677 0.936,0.271 2.136,-0.227 1.272,-1.369 -0.487,-0.645 -2.671,-1.477 -1.375,-2.397 -0.584,-0.013 -1.16,-0.085 -1.729,-0.216 0.178,-0.489 0.248,-0.993 0.21,-1.512 -0.661,0.323 -1.571,-0.455 -1.938,-0.916 -0.686,-0.861 -1.271,-0.255 -2.232,-0.668 0.334,0.769 0.756,1.5 1.01,2.304 -0.307,-0.495 -0.597,-0.999 -0.87,-1.512 -0.13,0.525 -0.033,1.006 0.29,1.44 -0.859,-1.18 -1.324,-1.92 -0.57,-3.313 -0.334,0.063 -0.67,0.086 -1.01,0.072 0.198,-0.511 0.411,-1.015 0.641,-1.512 -0.14,0.031 -0.257,0.104 -0.351,0.216 -0.146,-0.552 -0.146,-1.104 0,-1.656 -0.5,1.209 -0.421,2.561 -1.01,3.744 -0.963,-0.422 -0.52,-0.948 -0.412,-1.705 0.126,-0.882 -0.887,-1.61 -1.317,-2.255 0.125,0.734 0.556,1.516 0.069,2.231 -0.361,-1.333 -1.739,-1.98 -2.503,-3.036 -0.926,-1.28 -1.477,-3.091 -2.174,-4.521"
-              id="path4608"
-              data-inkscape-connector-curvature="0"
-            />
-          </g>
-          <g
-            transform="matrix(6.4876122,0,0,6.4876122,-12925.057,-3966.0743)"
-            style={{
-              fill: clickedCountryId === "Vietnam" ? "#FF0000" : "#005500",
-              fillOpacity: 0.98823529,
-              stroke: "#ffffff",
-              strokeWidth: 0.5257113,
-              strokeMiterlimit: 3.98288536,
-              strokeDasharray: "none",
-              strokeOpacity: 1,
-            }}
-            id="Vietnam"
-            className="country"
-          >
-            <path
-              d="m 2108.088,720.288 c 0.238,0.767 0.688,1.553 1.51,1.8 -0.58,-0.527 -1.108,-1.123 -1.51,-1.8"
-              style={{
-                fill: "#FFC0CB",
-                fillOpacity: 0.98823529,
-                stroke: "#ffffff",
-                strokeWidth: 0.5257113,
-                strokeMiterlimit: 3.98288536,
-                strokeDasharray: "none",
-                strokeOpacity: 1,
-              }}
-              id="path3474"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2112.698,712.368 c 0.168,0.788 -0.383,1.58 0.43,2.16 0.366,-0.76 0.116,-1.568 -0.43,-2.16"
-              id="path3476"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2113.918,713.664 c -0.059,0.238 -0.059,0.478 0,0.72 0,-0.24 0,-0.48 0,-0.72"
-              id="path3478"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2130.559,695.304 c 0.117,0.339 0.354,0.555 0.709,0.648 -0.237,-0.216 -0.473,-0.432 -0.709,-0.648"
-              id="path3480"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2065.538,610.848 c 1.348,0.063 1.071,-1.897 1.578,-2.671 0.825,-1.26 3.072,0.699 3.705,1.333 1.132,1.133 1.959,0.2 2.657,-0.994 0.96,-1.642 1.402,-0.101 2.14,0.748 0.274,-0.635 0.26,-2.187 1.254,-1.194 0.614,0.613 1.182,1.368 1.923,1.835 0.463,0.292 0.782,-2.749 2.083,-2.801 -0.372,2.019 3.378,0.194 3.985,-0.421 0.752,-0.76 -0.141,-1.278 0.023,-1.942 0.267,-1.082 2.592,-1.189 3.111,-2.388 1.066,1.475 2.415,2.638 4.047,3.425 1.277,0.617 2.453,-0.563 3.377,0.567 0.927,1.134 2.364,-0.626 3.492,0.502 0.572,0.573 1.268,2.896 -0.326,2.346 -0.01,1.967 1.037,3.412 1.65,5.327 0.569,-0.489 2.67,0.176 2.67,1.008 0,1.247 2.4,0.968 2.73,2.161 0.24,-0.144 0.48,-0.288 0.72,-0.433 0.388,1.019 1.069,0.258 1.915,0.002 1.061,-0.321 1.949,0.872 2.694,1.438 -0.864,0.757 -1.057,-0.684 -1.942,0.051 -0.802,0.666 -1.059,1.615 -2.14,1.783 -1.158,0.18 -0.246,1.235 -0.67,2 -0.413,0.745 -2.751,1.466 -1.577,0.126 -0.754,-0.183 -1.762,0.964 -1.801,1.368 -0.635,-0.355 -0.883,-1.538 -1.8,-1.225 0.284,0.308 0.617,0.547 1,0.72 -0.636,0.744 -0.156,1.112 0.326,1.902 0.402,0.658 -1.051,0.639 -1.326,0.474 0.327,0.681 0.51,2.588 -0.22,3.097 0.097,0.032 0.193,0.032 0.29,0 -0.078,0.545 -2.325,2.995 -2.811,2.376 -1.212,0.605 -1.601,2.421 -1.718,3.629 -0.162,1.667 0.541,2.796 -0.321,4.364 -0.978,1.778 0.92,2.184 1.029,3.814 0.429,-1.125 0.685,0.571 1.15,1.184 0.626,0.824 1.354,1.463 2.2,2.053 0.805,0.562 1.619,1.146 2.338,1.817 1.033,0.964 0.58,2.001 0.932,3.227 -0.444,-0.346 -0.947,-0.538 -1.51,-0.576 1.798,0.305 2.647,2.548 4.053,3.674 1.026,0.822 4.278,2.423 3.438,4.174 1.1,-0.635 2.849,1.567 3.459,2.232 -0.306,-0.14 -0.619,-0.26 -0.939,-0.36 1.345,0.838 2.573,1.759 3.46,3.097 1.198,-1.436 3.618,1.785 2.95,2.52 0.634,-0.444 -0.348,-1.832 0.939,-1.152 -1.93,1.663 3.992,5.51 4.101,6.984 0.491,-1.251 0.718,-0.207 1.112,0.368 0.616,0.899 -0.329,0.53 -0.173,1.216 0.29,1.276 1.326,2.412 1.833,3.603 0.687,1.614 1.045,3.365 1.772,4.987 0.293,0.653 0.608,1.313 0.768,2.017 0.078,0.344 0.11,2.158 -0.343,0.409 -0.026,0.975 1.122,3.806 0,2.089 0.255,0.579 0.614,1.083 1.08,1.512 -1.34,0.184 -0.261,0.854 -0.049,1.614 0.307,1.1 0.656,1.843 1.24,2.82 0.238,0.398 0.253,2.376 0.168,2.981 -0.406,-0.409 -0.693,-0.889 -0.859,-1.44 -1.36,0.797 -0.084,2.447 0.359,3.385 -0.45,-0.266 -0.931,-0.458 -1.439,-0.576 0.644,0.585 1.297,3.662 0.939,4.535 -0.01,-0.521 -0.149,-1 -0.43,-1.439 0.451,0.716 0.068,1.457 -0.149,2.16 0.812,-0.022 0.854,0.57 0.641,1.186 -0.228,0.657 -1.189,0.971 -1.251,1.19 -0.377,1.351 0.152,1.39 -1.32,2.145 -0.414,0.355 -0.852,0.677 -1.313,0.965 -0.641,0.262 -0.843,1.19 -1.333,1.338 -0.812,0.245 -1.648,0.592 -2.359,1.058 -0.333,0.218 -0.835,1.832 -1.091,1.745 -0.731,-0.249 -1.959,0.658 -2.512,1.075 -0.75,0.566 -2.232,1.064 -2.781,1.755 -0.509,-0.988 -1.846,-0.956 -1.94,-2.232 -0.095,0.505 -0.19,1.009 -0.289,1.513 -0.115,-0.406 -0.282,-0.79 -0.5,-1.152 0.373,0.833 0.249,0.785 -0.371,-0.145 0.811,0 -0.806,-0.508 -0.789,-0.504 0.248,0.874 -0.106,1.994 -1.221,1.729 0.263,0.805 1.422,-0.417 1.688,0.72 0.34,1.451 -2.248,0.74 -2.768,0.72 0.897,0.728 2.371,0.587 2.95,1.729 -0.869,0.132 -0.918,1.503 -1.916,0.805 -0.662,-0.463 -1.101,-1.856 -1.974,-1.957 0.559,0.655 3.338,2.954 2.88,3.672 -1.772,-0.721 -2.702,-2.48 -4.17,-3.6 0.316,1.153 5.591,4.695 3.225,5.987 -1.554,0.849 -4.187,-3.283 -5.314,-3.972 0.768,1.263 3.607,3.212 3.122,4.893 -0.376,1.3 -2.584,1.655 -3.629,2.178 -1.764,0.883 -1.824,2.202 -2.904,3.611 -0.735,0.961 -2.648,2.555 -3.859,1.414 0.371,-0.47 0.852,-0.758 1.44,-0.864 -0.135,-0.3 -0.231,-0.612 -0.29,-0.937 -0.305,0.142 -0.618,0.262 -0.94,0.36 0.048,-1.815 -0.328,-3.617 -0.341,-5.433 -0.01,-0.895 0.058,-1.812 0.341,-2.667 0.19,-0.576 1.464,-3.022 1.73,-1.261 -0.01,-2.827 -5.396,-2.415 -4.89,-4.823 -0.104,0.426 -0.297,0.402 -0.58,-0.072 0.482,-0.166 0.781,-0.937 1.323,-0.937 0.576,0 1.222,0.156 1.771,-0.105 0.46,-0.22 0.976,-0.896 1.283,-1.293 0.595,-0.764 -0.491,-1.525 0.232,-2.273 1.366,1.314 3.278,-0.11 4.753,-0.4 0.513,-0.101 0.805,0.346 1.037,0.725 0.562,0.915 1.663,0.828 2.56,1.116 0.021,-0.788 -0.803,-1.446 0.141,-1.584 -0.183,-1.33 -1.845,-1.38 -2.389,-2.302 -0.466,-0.788 -0.778,-1.8 -0.968,-2.692 -0.181,-0.846 0.708,-0.943 1.129,-1.597 0.642,-0.996 2.779,0.455 3.598,0.614 -0.457,-0.836 0.135,-1.823 -0.5,-2.592 0.695,-0.175 1.464,0.183 2.145,-0.113 0.577,-0.251 1.739,-1.017 2.297,-1.406 0.647,-0.453 0.786,-1.099 1.607,-1.334 0.635,-0.181 1.488,-0.11 2,-0.593 1.379,-1.301 -0.729,-4.871 -0.554,-6.536 0.148,-1.427 0.679,-2.45 -0.027,-3.83 -0.232,-0.423 -0.467,-0.845 -0.704,-1.266 -0.4,-0.753 -0.15,-1.289 -0.434,-1.985 -0.401,-0.984 -0.953,-0.894 -1.051,-2.205 -0.118,-1.586 0.892,-2.514 0.984,-3.918 0.022,-0.338 0.027,-2.105 -0.166,-2.266 -1.2,-0.999 -0.057,-0.752 0.397,-1.796 0.7,-1.609 -1.21,-3.196 -2.367,-3.942 -0.53,-0.342 -1.896,-1.068 -1.828,-1.856 0.081,-0.941 1.652,-1.326 2.02,-2.16 -1.006,-0.382 -2.312,-1.165 -2.955,-2.051 -0.479,-0.658 -0.928,-0.111 -1.458,-1.013 -0.715,-1.215 -1.111,-0.084 -1.996,-0.608 -0.515,-0.306 -1.105,-1.607 -1.203,-2.185 -0.308,-1.819 -1.064,-3.372 -2.552,-4.654 -0.42,-0.362 -1.376,-0.603 -1.964,-1.22 -1.107,-1.165 -2.436,-2.346 -3.371,-3.649 -0.602,-0.838 -0.86,-1.866 -1.491,-2.684 -0.251,-0.324 -0.67,-0.711 -1.09,-0.369 -0.441,-0.678 -1.261,-0.915 -1.678,-1.613 -0.307,-0.514 -1.177,-1.727 -0.598,-2.301 0.724,-0.719 -2.672,-1.417 -2.897,-1.469 -2.263,-0.517 -3.809,-3.112 -5.952,-3.7 -0.619,-0.17 -2.291,-0.956 -0.947,-1.588 1.91,-0.898 -0.494,-1.428 -0.157,-2.432 0.18,-0.54 2.777,0.254 3.034,0.293 1.056,0.16 3.916,-1.596 2.436,-2.021 0.323,-0.489 1.29,-0.747 1.29,-1.44 0,-0.875 -0.971,-1.66 -1.821,-1.549 -0.578,0.075 -0.648,-0.299 -0.816,-0.782 -0.371,-1.067 -1.475,-0.779 -2.322,-1.125 0.197,-0.759 1.301,-0.61 1.361,-1.241 0.104,-1.071 -1.08,-1.476 -1.776,-1.98 -0.831,-0.602 -1.391,-1.325 -2.555,-1.113 -1.813,0.329 -1.409,1.298 -2.22,2.391 -0.806,-0.494 -1.994,-1.428 -2.878,-1.257 -1.008,0.195 -2.89,-1.614 -3.382,-2.344 -0.561,-0.829 -1.07,-2.813 -0.785,-3.805 0.504,-1.757 -0.951,-1.373 -1.306,-2.747 0.221,0.564 0.124,1.068 -0.29,1.512 -0.945,-0.286 -1.158,-1.962 -1.975,-2.608 -1.104,-0.881 -2.922,-2.432 -3.494,-3.733"
-              id="path3562"
-              className="Vietnam"
-              data-inkscape-connector-curvature="0"
-            />
-
-            <path
-              d="m 2109.748,619.632 c 0.25,0.118 0.49,0.07 0.72,-0.144 -0.24,0.048 -0.48,0.096 -0.72,0.144"
-              id="path5502"
-              data-inkscape-connector-curvature="0"
-            />
-          </g>
-          <g
-            transform="matrix(6.4876122,0,0,6.4876122,-12731.067,-3673.6483)"
-            style={{
-              fill: clickedCountryId === "Malaysia" ? "#FF0000" : "#005500",
-              fillOpacity: 0.98823529,
-              stroke: "#ffffff",
-              strokeWidth: 0.5257113,
-              strokeMiterlimit: 3.98288536,
-              strokeDasharray: "none",
-              strokeOpacity: 1,
-            }}
-            id="Malaysia"
-            className="country"
-          >
-            <path
-              d="m 2034.2663,703.94153 c 0.396,-0.539 0.328,-1.529 0.591,-2.153 0.459,-1.091 1.521,1.21 2.075,1.534 0.688,0.402 1.312,0.133 2.02,0.331 0.664,0.186 0.708,0.962 1.055,1.44 0.686,0.946 2.455,0.748 1.946,2.594 -0.246,0.891 -1.392,3.415 0.884,2.878 1.021,-0.24 1.24,-1.943 2.425,-2.13 0.986,-0.156 1.402,1.773 2.315,1.417 0.741,-0.289 1.839,-3.471 1.81,-4.254 0.92,0.292 1.767,0.383 2.38,1.224 0.607,0.833 0.848,1.846 1.684,2.515 2.081,1.665 3.911,3.196 5.147,5.601 1.249,2.428 2.123,4.887 1.986,7.668 -0.053,1.069 -0.193,2.252 -0.397,3.303 -0.289,1.489 -0.02,1.601 0.429,2.94 0.669,2.003 -0.554,4.433 0.637,6.448 0.92,1.558 3.025,2.387 3.529,4.274 0.591,2.231 2.061,4.105 2.626,6.363 0.759,3.031 -2.675,1.37 -1.101,0.271 -0.603,0.022 -1.117,-0.327 -1.37,-0.864 0.107,0.208 0.671,2.235 -0.165,1.632 -0.841,-0.608 -2.808,0.733 -3.575,1.104 -0.81,-2.641 -4.021,-3.369 -5.938,-4.992 -0.495,-0.419 -0.871,-0.924 -1.344,-1.36 -0.757,-0.699 -1.961,-0.68 -2.798,-1.281 -0.842,-0.604 -2.181,-1.896 -2.646,-2.768 -0.321,-0.603 -0.819,-0.66 -1.417,-0.821 -0.928,-0.25 -1.345,-1.086 -2.091,-1.607 -0.711,-0.497 -0.563,-0.633 -0.178,-1.175 0.429,-0.604 -0.136,-1.837 -0.387,-2.404 -0.788,-1.778 -2.784,-3.004 -3.507,-4.803 -0.349,-0.866 -2.242,-1.346 -0.075,-2.046 -1.277,0.183 -2.556,-2.476 -2.217,-3.412 0.657,-1.816 -0.308,-3.773 -1.403,-5.126 -0.965,-1.19 -0.765,-4.93 -0.77,-6.438 -0.01,-2.205 -1.347,-3.95 -2.16,-5.903"
-              id="path4722"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2169.1963,723.52553 c -0.047,-0.264 -0.093,-0.528 -0.14,-0.792 0.696,-0.06 1.261,0.352 1.8,0.721 -0.556,-0.024 -1.109,0 -1.66,0.071"
-              id="path2560"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2120.9563,737.92553 c 0.069,0.498 0.159,0.74 0.72,0.792 -0.389,-0.997 0.126,-2.281 -0.86,-3.024 -0.049,0.752 0.186,1.481 0.14,2.232"
-              id="path4708"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2149.4663,713.87753 c 0.733,0.266 0.829,0.025 0.29,-0.721 -0.096,0.241 -0.193,0.481 -0.29,0.721"
-              id="path4710"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2161.5663,697.38953 c 0.431,-0.246 0.717,-0.605 0.86,-1.08 -0.566,0.095 -0.959,0.473 -0.86,1.08"
-              id="path4712"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2163.2863,696.45353 c -0.245,0.472 -0.219,0.927 0.08,1.367 1.543,-0.537 0.909,-2.4 -0.08,-1.367"
-              id="path4714"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2165.4463,701.85353 c 0.288,0.052 0.555,-0.02 0.801,-0.216 -0.267,0.072 -0.535,0.143 -0.801,0.216"
-              id="path4716"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2174.3763,719.13353 c 0.417,0.44 0.896,0.512 1.439,0.216 -0.473,-0.111 -0.953,-0.182 -1.439,-0.216"
-              id="path4718"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2078.0463,734.97353 c 0.706,-0.407 0.751,-1.327 -0.15,-1.512 0.136,0.495 0.186,0.999 0.15,1.512"
-              id="path2284"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2101.5863,734.46953 c 0.249,0.201 0.513,0.201 0.79,0 -0.074,-0.263 -0.194,-0.503 -0.36,-0.72 -0.144,0.24 -0.286,0.48 -0.43,0.72"
-              id="path2300"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2032.1763,703.65353 c -0.352,0.384 -0.758,0.696 -1.22,0.937 0.773,1.484 2.401,0.017 1.22,-0.937"
-              id="path4700"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2036.0663,712.43753 c -0.876,-0.385 -0.966,0.444 -0.914,1.081 0.109,1.343 1.077,-0.2 0.914,-1.081"
-              id="path4704"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2066.4463,734.75753 c -0.292,0.177 -0.436,0.44 -0.43,0.792 0.828,0.15 0.972,-0.114 0.43,-0.792"
-              id="path4706"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2108.3563,741.59753 c -0.245,2.565 5.35,4.75 5.26,2.376 0.027,0.67 1.017,0.744 0.95,1.021 -0.201,0.847 1.416,1.213 2.013,1.433 0.921,0.339 1.471,-0.408 2.26,0.495 0.719,0.824 1.81,0.915 2.838,1.011 -0.716,-0.983 -2.25,-0.124 -2.628,-1.594 -0.244,-0.953 0.498,-1.752 1.328,-0.925 -0.796,-0.499 -0.847,-1.277 -0.581,-2.097 0.167,-0.395 0.293,-0.802 0.379,-1.222 0.213,-1.005 0.741,-0.613 1.352,-1.218 -0.503,0.053 -1.006,0.101 -1.51,0.144 0.694,-0.664 0.016,-1.451 0.221,-2.231 0.75,0.006 1.415,0.393 2.159,0.432 -2.607,-1.956 1.313,-4.282 2.925,-4.538 2.552,-0.405 6.229,-0.939 8.382,-2.456 1.86,-1.31 1.92,-4.002 3.57,-5.487 2.114,-1.902 3.326,-4.094 3.483,-6.959 1.507,0.045 1.805,0.81 2.355,2.019 0.323,0.71 2.139,2.679 2.918,2.712 1.381,0.059 1.772,-2.784 1.188,-3.608 -1.316,-1.858 0.31,-2.611 1.308,-3.715 0.348,1.824 -0.03,4.37 2.311,4.896 0.108,-1.41 -0.227,-3.004 -0.955,-4.218 -0.932,-1.552 0.832,-0.637 1.392,-1.084 0.517,-0.413 2.312,-1.796 0.884,-2.351 -2.156,-0.838 -0.048,-2.439 0.619,-3.364 -0.178,1.821 2.072,-0.156 2.47,-0.934 0.607,-1.188 0.8,-2.535 1.357,-3.744 0.317,-0.689 0.713,-0.355 1.073,-0.794 0.437,-0.531 0.806,-1.433 1.268,-2.053 1.011,-1.355 0.895,-3.324 2.192,-4.571 1.008,1.186 -0.201,2.409 0.28,3.672 0.824,-0.524 1.585,-3.679 2.604,-3.493 0.069,0.013 2.517,5.6 2.875,3.062 0.128,1.043 2.063,1.501 1.922,2.753 -0.093,0.827 -1.245,0.523 -0.948,1.653 0.348,1.322 0.178,1.66 -1.123,2.146 0.812,-0.466 1.682,0.053 2.426,-0.372 0.608,-0.347 0.937,-1.22 1.795,-0.859 1.623,0.682 0.105,1.806 -0.691,2.095 0.165,1.559 1.67,0.065 2.265,-0.313 1.277,-0.812 2.789,1.076 2.916,2.185 0.881,-0.934 2.38,0.719 2.739,1.439 0.708,-0.968 2.712,-0.218 2.74,0.932 0.042,1.661 -3.604,2.974 -4.915,2.872 -0.911,-0.071 -4.406,-1.292 -3.327,1.257 0.425,1.004 2.745,1.429 2.553,2.644 0.71,-0.445 1.279,0.219 0.867,0.89 -0.442,0.723 -1.447,0.564 -2.153,0.622 -2.26,0.182 -3.2,0.76 -5.124,-0.863 -0.167,0.518 0.193,1.6 -0.325,1.959 -0.785,0.544 -1.896,-0.578 -2.478,-0.958 -1.729,-1.128 -4.083,0.636 -5.627,-0.857 -0.109,1.731 -1.258,0.146 -2.096,0.376 -0.623,0.171 -0.814,-0.417 -1.255,0.292 -0.491,0.793 -1.253,-0.31 -1.83,-0.524 -0.168,0.77 -0.811,1.008 -1.149,1.655 -0.438,0.838 -0.66,2.005 -0.775,2.938 -0.091,0.738 -0.196,1.613 0.027,2.34 0.347,1.132 -0.072,1.269 -0.313,2.329 -0.387,1.696 0.123,2.488 -1.813,2.528 -2.099,0.043 -0.81,3.689 -0.226,4.411 -1.166,0.403 -4.292,2.373 -2.898,3.917 0.809,0.895 -2.133,3.913 -1.852,5.299 -1.201,0.061 -1.211,-0.777 -2.453,0.009 -0.856,0.542 -1.728,-0.311 -2.569,0.143 -1.148,0.619 -1.44,1.63 -2.974,1.384 -1.012,-0.163 -3.096,-0.874 -3.643,-1.809 -0.745,-1.272 -5.502,-1.106 -6.348,0.417 -0.399,0.717 -0.634,1.614 -0.896,2.385 0.054,-0.158 -2.323,1.254 -2.221,1.246 -0.698,0.054 -1.396,0.154 -2.09,0.25 -0.825,0.113 -1.519,-0.52 -2.379,-0.574 -1.491,-0.094 -2.461,0.696 -3.728,1.3 -1.736,0.828 -1.795,-0.581 -3.209,-0.951 -0.435,-0.113 -1.413,-1.269 -1.831,-1.641 -0.524,-0.468 -0.736,-1.158 -1.25,-1.628 -0.505,-0.463 -1.155,-0.421 -1.499,-1.093 -0.314,-0.615 -0.143,-2.145 -0.793,-2.506 -1.328,-0.739 -0.149,-1.262 0.011,-2.335"
-              id="path4724"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2182.5163,719.42153 c 0.381,-0.749 0.309,-1.573 -0.07,-2.304 0.087,0.766 0.07,1.536 0.07,2.304"
-              id="path5822"
-              data-inkscape-connector-curvature="0"
-            />
-          </g>
-          <g
-            transform="matrix(6.4876122,0,0,6.4876122,-12925.057,-3966.0743)"
-            style={{
-              fill: clickedCountryId === "Philippines" ? "#FF0000" : "#005500",
-              fillOpacity: 0.98823529,
-              stroke: "#ffffff",
-              strokeWidth: 0.5257113,
-              strokeMiterlimit: 3.98288536,
-              strokeDasharray: "none",
-              strokeOpacity: 1,
-            }}
-            id="Philippines"
-            className="country"
-          >
-            <path
-              d="m 2249.568,695.88 c 0.197,0.178 0.341,0.153 0.43,-0.072 -0.144,0.024 -0.286,0.047 -0.43,0.072"
-              id="path5902"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2192.328,737.136 c 0.481,-0.783 0.485,-1.705 -0.29,-2.304 -0.323,0.773 -0.221,1.636 0.29,2.304"
-              id="path5800"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2204.348,744.336 c 0.094,-0.796 -0.193,-0.964 -0.859,-0.504 0.269,0.196 0.556,0.364 0.859,0.504"
-              id="path5806"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2215.298,760.968 c 0.45,-0.254 3.06,-1.666 2.938,-1.991 -0.446,-1.199 -1.203,-0.073 -1.813,0.361 -0.749,0.537 -1.981,0.559 -1.125,1.63"
-              id="path5832"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2219.979,749.52 c -0.169,0.397 -0.146,0.781 0.069,1.152 0.349,-0.414 0.326,-0.798 -0.069,-1.152"
-              id="path5838"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2225.878,753.983 c 1.974,-0.984 -1.108,-1.938 -2.054,-1.763 -1.029,0.192 -0.889,1.498 -0.648,1.368 0.668,-0.359 2.399,-0.526 2.702,0.395"
-              id="path5840"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2229.988,748.008 c 0.141,0.87 0.968,1.36 1.785,0.986 0.835,-0.38 0.777,-1.284 1.305,-1.922 -0.741,0.063 -1.188,-0.512 -1.87,-0.648 -0.704,-0.142 -1.233,0.725 -1.87,0.864 0.19,0.264 0.407,0.504 0.65,0.72"
-              id="path5858"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2229.408,752.112 c 0.21,0.558 0.569,0.726 1.08,0.504 -0.328,-0.237 -0.688,-0.405 -1.08,-0.504"
-              id="path5860"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2236.968,741.672 c 2.662,-1.05 -2.155,-1.712 0,0"
-              id="path5874"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2241.578,725.976 c 1.243,-0.064 0.718,-1.202 0.22,-1.8 -0.702,0.346 -1.136,1.382 -0.22,1.8"
-              id="path5882"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2242.588,722.088 c 0.684,-0.043 0.924,-0.283 0.721,-0.72 -0.561,-0.045 -0.746,0.185 -0.721,0.72"
-              id="path5888"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2243.598,716.04 c 0.19,-0.24 0.38,-0.479 0.57,-0.72 -0.668,0.011 -0.573,0.195 -0.57,0.72"
-              id="path5892"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2250.648,726.12 c 0.341,-0.751 -0.293,-1.918 -1.15,-1.152 0.161,0.556 0.587,0.999 1.15,1.152"
-              id="path5900"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2256.908,758.088 c 0.231,0.131 0.475,0.155 0.729,0.072 -0.241,-0.229 -0.484,-0.253 -0.729,-0.072"
-              id="path5908"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2259.508,745.056 c 0.227,-0.776 0.012,-1.544 -0.149,-2.304 -1.248,0.25 -0.166,1.494 0.149,2.304"
-              id="path5910"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2193.698,730.368 c -0.386,0.693 -0.252,1.49 -0.43,2.231 1.96,-2.429 4.747,-3.229 6.479,-5.976 0.645,-1.021 6.15,-4.616 4.17,-6.264 1.169,0.289 0.37,-0.516 0.441,-1.225 0.087,-0.864 1.477,-0.983 2.08,-1.063 1.033,-0.136 1.234,-0.538 1.405,-1.487 0.239,-1.323 1.195,-0.994 2.054,-1.554 1.808,-1.176 0.479,-2.33 -0.278,-3.686 -0.909,-1.628 0.562,-3.855 -0.802,-5.243 -0.876,0.764 0.194,3.678 -1.08,2.521 0.313,0.602 1.152,2.289 1.22,2.952 -0.739,-0.412 -1.594,-0.857 -1.87,-1.729 -0.015,0.913 0.459,1.645 0.737,2.478 0.28,0.837 -0.775,3.423 -1.237,1.627 -0.069,0.558 -0.372,0.742 -0.875,0.88 -0.389,0.106 -0.178,1.22 -0.285,1.496 -0.135,0.348 -0.842,-0.104 -1,0.216 -0.252,0.508 0.417,0.86 0,1.439 -0.189,-1.695 -0.907,0.266 -1.16,0.723 -0.421,0.762 -0.823,1.536 -1.411,2.188 -0.931,1.032 -1.416,2.441 -2.506,3.342 -0.434,0.358 -0.928,0.192 -1.361,0.515 -0.496,0.369 -0.523,1.018 -1.035,1.317 -1.452,0.852 -2.697,2.725 -3.256,4.302"
-              id="path5802"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2211.338,706.176 c 0.208,-0.762 -0.055,-0.905 -0.79,-0.432 0.264,0.144 0.526,0.288 0.79,0.432"
-              id="path5812"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2211.838,698.544 c -0.218,-0.219 -0.458,-0.411 -0.72,-0.576 -1.27,2.401 1.873,2.809 3.53,2.809 0.048,-0.442 -0.213,-1.039 -0.611,-1.26 -0.82,-0.457 -1.338,-0.258 -2.199,-0.973"
-              id="path5814"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2211.048,684.864 c 0.282,0.695 0.997,1.122 1.59,1.512 0.034,-0.89 -0.768,-1.477 -1.59,-1.512"
-              id="path5816"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2212.128,701.424 c -0.313,0.072 -0.626,0.144 -0.939,0.216 0.172,0.892 1.242,0.966 1.159,1.944 0.653,-0.589 -0.011,-1.566 -0.22,-2.16"
-              id="path5818"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2213.278,713.376 c -0.097,-1.258 -1.732,-0.092 -1.72,0.72 0.97,0.388 0.797,-0.851 1.72,-0.72"
-              id="path5820"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2213.138,688.176 c 0.154,0.414 1.751,1.692 2.13,1.991 1.613,1.274 1.217,2.934 2.389,4.274 0.514,0.588 3.993,6.938 4.767,3.239 0.164,-0.784 0.644,-1.448 0.796,-2.232 0.15,-0.778 -0.581,-1.471 -0.581,-2.25 0,-1.86 -0.938,-3.017 -2.196,-4.271 -1.015,-1.017 -6.339,-1.998 -7.305,-0.751"
-              id="path5824"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2214.288,702.504 c -0.024,-0.511 -0.099,-1.015 -0.22,-1.512 -0.277,0.558 -0.204,1.062 0.22,1.512"
-              id="path5828"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2206.229,665.063 c -0.323,1.593 0.957,1.391 1.116,2.294 0.243,1.383 0.348,3.162 1.043,4.381 0.67,1.174 1.017,2.204 1.226,3.516 0.127,0.795 1.399,2.547 1.225,0.753 1.342,0.504 0.572,3.118 2.127,3.584 2.294,0.688 0.481,-2.257 0.533,-3.224 0.581,0.212 5.333,1.771 3.49,3.086 -0.707,0.504 -1.327,1.115 -1.964,1.701 -1.021,0.939 0.021,2.794 0.124,3.997 0.612,-0.073 0.852,-0.437 0.58,-1.009 1.296,0.021 1.89,0.74 1.659,2.017 0.422,-1.496 1.363,-0.183 1.08,0.576 0.738,-0.122 1.825,0.334 2.436,-0.277 0.521,-0.522 0.666,-1.318 1.261,-1.767 0.185,-0.139 1.266,-0.97 1.53,-0.759 0.904,0.724 2.11,1.083 2.941,1.875 1.099,1.046 3.193,2.474 3.493,4.023 0.1,0.519 1.013,1.276 1.219,0.366 -0.011,-0.491 -0.146,-0.95 -0.401,-1.379 -0.1,-0.241 -1.172,-1.691 -0.599,-1.508 -1.151,-0.723 -0.851,-2.273 -1.659,-3.239 1.554,-0.246 3.011,1.626 3.759,2.695 1.013,1.446 2.738,1.532 3.605,3.314 0.227,0.468 -0.297,1.374 0.386,1.521 0.741,0.159 1.364,0.686 2.043,1.012 0.542,0.261 1.014,1.015 1.634,0.688 0.451,-0.237 1.12,-0.925 1.683,-0.807 -0.115,0.884 -1.742,1.279 -1.035,2.308 0.128,0.186 1.472,2.115 1.829,1.163 0.555,-1.479 -0.043,-6.612 -2.813,-4.694 -0.31,-0.609 0.011,-1.016 -0.134,-1.61 -0.184,-0.752 -1.271,-1.835 -1.916,-2.207 -1.711,-0.986 1.9,-1.95 2.84,-1.44 -0.163,-1.06 -1.879,-1.486 -2.745,-1.565 -1.476,-0.136 -1.724,-1.385 -2.945,-1.603 -0.307,1.002 1.567,2.461 0,3.024 -1.147,0.412 -1.485,-2.342 -1.992,-3.2 -1.031,-1.75 -5.363,-3.313 -5.277,-0.04 -0.175,-0.298 -0.365,-0.586 -0.57,-0.864 -0.972,0.9 1.006,1.464 0.45,2.095 -0.939,1.068 -2.865,-0.383 -3.55,-1.086 -0.614,-0.631 -3.257,-5.141 -1.301,-5.257 -1.774,-1.237 -3.611,-4.797 -2.8,-6.84 0.553,-1.393 -0.152,-2.662 0.351,-3.888 0.621,-1.518 2.652,-1.619 3.68,-2.736 -0.118,0.409 -0.238,0.817 -0.36,1.225 0.867,-0.986 0.593,-2.143 0.995,-3.311 0.614,-1.78 1.432,-3.69 1.095,-5.617 -1,0.473 -0.544,-0.909 -0.869,-1.369 -0.9,-1.271 -1.942,-1.312 -2.494,-3.061 -0.564,-1.795 -0.519,-3.185 0.23,-4.872 0.389,-0.875 -0.746,-3.274 -1.534,-1.997 -0.313,0.507 -0.369,1.625 -1.228,1.469 -0.52,-0.095 -1.643,-0.693 -2.105,-0.106 -0.277,-0.601 -0.856,-1.025 -1.399,-1.362 -1.089,-0.675 -2.507,-1.688 -3.798,-1.26 -0.557,0.185 -1.163,-0.118 -1.717,-0.167 -0.517,-0.047 -0.118,0.701 -0.605,0.701 -1.385,0 -0.513,0.765 -0.54,1.584 -0.05,1.464 -0.508,3.266 -0.72,4.752 -0.387,2.704 1.655,5.093 1.092,7.969 -0.456,2.332 0.506,4.072 0.819,6.243 0.39,2.691 -3.71,-0.898 -4.001,-2.044 -0.908,0.455 -0.82,1.436 -0.503,2.233"
-              id="path5808"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2213.718,637.128 c 0.331,0.49 0.784,0.802 1.36,0.936 0.213,-0.815 -0.685,-1.375 -1.36,-0.936"
-              id="path5826"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2214.288,641.592 c 0.264,-0.024 0.526,-0.048 0.79,-0.072 -0.412,-0.183 -0.602,-0.319 -0.79,0.072"
-              id="path5830"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2216.738,628.632 c 0.232,-0.352 0.375,-0.735 0.43,-1.152 -0.47,0.261 -0.613,0.645 -0.43,1.152"
-              id="path5834"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2218.248,641.808 c 0.479,-0.356 0.698,-0.935 0.51,-1.512 -0.445,0.413 -0.615,0.917 -0.51,1.512"
-              id="path5836"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2222.928,674.856 c 0.235,0.884 0.696,2.579 1.66,2.951 0.423,-1.08 0.119,-4.529 -1.66,-2.951"
-              id="path5842"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2226.678,683.496 c -0.216,-0.876 -1.382,-1.484 -2.159,-1.729 0.495,0.82 1.379,1.24 2.159,1.729"
-              id="path5844"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2224.519,688.32 c -0.027,1.616 2.487,2.827 2.56,0.517 0.035,-1.176 -2.184,-1.815 -2.56,-0.517"
-              id="path5846"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2226.458,677.304 c 0.215,-0.694 -0.025,-0.886 -0.72,-0.576 0.241,0.192 0.48,0.383 0.72,0.576"
-              id="path5848"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2226.388,702.792 c 2.336,0.414 1.775,2.229 1.731,4.019 -0.031,1.298 0.182,2.221 -0.168,3.52 -0.363,1.354 0.424,2.644 -0.053,3.981 1.34,0.284 2.196,-2.016 3.657,-2.016 0.951,0 2.723,-1.387 2.447,-2.333 -0.241,-0.828 2.796,-1.137 2.386,-2.317 -0.545,-1.565 0.57,-3.722 -1.94,-1.901 0.616,-0.673 -0.31,-1.585 -1.047,-1.421 -0.684,0.152 -1.129,1.013 -1.715,0.39 -0.663,-0.704 -1.135,-1.509 -2.046,-1.938 -0.857,-0.403 -1.965,-0.573 -2.532,-1.423 -0.379,0.409 -0.619,0.889 -0.72,1.439"
-              id="path5850"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2226.958,696.456 c -0.099,1.3 -0.635,2.405 0.36,3.527 0.176,-1.48 0.295,-2.963 0.604,-4.41 0.174,-0.818 -1.671,-0.3 -0.964,0.883"
-              id="path5852"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2227.038,677.664 c 0.24,-0.024 0.479,-0.048 0.72,-0.072 -0.249,-0.042 -0.488,-0.018 -0.72,0.072"
-              id="path5854"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2228.908,696.456 c 0.624,0.094 0.551,-0.146 -0.22,-0.72 0.074,0.241 0.147,0.48 0.22,0.72"
-              id="path5856"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2229.988,742.752 c -0.371,2.035 2.42,2.893 2.632,0.514 0.077,-0.866 0.452,-3.26 1.114,-3.524 0.527,-0.209 0.906,-3.364 2.107,-2.436 0.864,0.668 0.51,2.104 1.196,2.854 0.19,-1.267 0.646,-0.651 1.01,0 0.282,-0.781 0.046,-1.861 1.011,-2.232 0.431,0.52 0.348,1.147 -0.15,1.584 0.991,0.036 1.614,1.261 2.6,1.08 -0.301,-0.436 -0.673,-1.028 -0.649,-1.584 0.013,-0.316 0.708,-0.28 0.79,-0.576 0.217,-0.779 -0.647,-0.597 0.357,-1.353 0.617,-0.464 1.226,0.352 1.693,0.645 1.589,0.996 3.024,1.271 3.557,3.377 0.288,1.145 -1.263,2.217 -1.385,3.44 -0.123,1.224 0.535,2.908 0.794,4.116 0.399,1.863 2.02,2.981 3.613,3.797 0.731,0.375 3.328,2.062 4.117,1.366 0.632,-0.557 0.297,-1.873 1.491,-1.811 0.118,0.006 -0.232,5.561 1.467,4.389 1.181,-0.814 2.207,-3.612 2.024,-4.976 -0.257,-1.918 -1.715,-2.722 -2.459,-4.27 -0.449,-0.935 1.825,-6.168 2.635,-5.78 0.623,0.298 2.002,3.261 2.274,3.982 0.266,0.702 0.114,1.435 0.176,2.161 0.09,1.06 0.949,1.767 1.104,2.796 0,-1.672 0.749,-3.84 -0.511,-5.256 0.621,-0.29 1.127,0.081 1.37,0.648 -1.069,-0.95 0.516,-1.902 1.037,-2.614 1.008,-1.375 0.299,-3.385 0.028,-4.856 -0.222,-1.208 -1.806,-1.213 -1.176,-2.525 0.365,-0.76 0.393,-2.717 -0.97,-2.101 2.318,-1.986 -1.923,-2.633 -1.869,-3.673 0.031,-0.608 1.214,-0.891 1.446,-1.488 0.416,-1.069 -0.774,-1.498 -1.087,-2.327 -0.147,-0.391 0.286,-2.259 -0.438,-1.885 -0.914,0.471 -1.083,0.138 -1.577,-0.569 -0.585,-0.838 0.01,-1.321 -1.274,-1.712 -1.41,-0.431 -1.666,-1.705 -2.899,-2.242 -0.745,1.625 0.565,3.763 1.056,5.298 0.111,0.351 0.623,1.645 -0.016,1.83 -0.833,0.24 -1.979,0.035 -2.341,-0.864 -0.677,2.121 0.048,2.292 -1.94,0.91 -1.871,-1.3 -0.246,3.828 -1.795,4.041 -0.995,0.137 -1.189,-1.064 -2.107,-0.651 -0.955,0.431 -0.691,2.183 -0.992,3 -0.361,0.982 -1.793,0.544 -2.506,1.124 -0.414,0.337 -1.127,1.134 -1.697,1.159 -0.588,0.026 0.73,-0.916 0.611,-0.836 0.745,-0.496 0.901,-0.772 0.713,-1.718 -0.146,-0.734 -1.153,-4.267 -2.376,-2.562 -0.187,-0.589 -0.548,-1.337 -1.3,-1.297 0.42,1.981 -2.057,1.221 -2.639,2.297 -0.688,1.27 -0.1,2.438 -1.682,2.959 -1.6,0.527 -3.966,0.067 -4.723,2.096 -0.324,0.87 -0.122,1.879 -0.222,2.783 -0.129,1.196 -0.751,2.391 -1.243,3.472"
-              id="path5862"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2230.848,696.456 c -0.262,0.164 -0.502,0.356 -0.72,0.576 0.613,0.487 1.242,0.959 1.8,1.512 0.676,-0.944 0.08,-2.094 -1.08,-2.088"
-              id="path5864"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2233.798,723.024 c 0.686,0.493 1.67,0.479 2.24,1.151 0.481,0.567 0.228,1.471 0.803,1.982 0.792,0.705 1.835,-0.137 2.231,-0.838 0.736,-1.305 -0.618,-2.194 -0.967,-3.334 -0.645,-2.099 0.169,-5.19 0.828,-7.236 0.501,-1.558 2.153,-3.65 -0.01,-4.791 -0.713,-0.376 -1.637,-0.495 -2.389,-0.155 -1.292,0.585 -0.751,1.712 -1.078,2.779 -0.211,0.688 -0.885,0.91 -0.635,1.729 0.278,0.91 0.671,1.904 0.306,2.857 -0.549,1.436 -3.195,0.518 -3.156,2.121 0.034,1.343 1.081,2.689 1.826,3.735"
-              id="path5866"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2232.718,714.6 c 0.644,-0.292 2.163,-2.283 0.795,-2.618 -1.033,-0.253 -1.298,2.082 -0.795,2.618"
-              id="path5868"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2237.038,694.8 c -0.828,-1.34 -1.782,-3.163 -3.31,-3.816 -0.753,1.266 2.74,2.988 3.31,3.816"
-              id="path5870"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2236.248,700.776 c -0.023,0.264 -0.047,0.527 -0.07,0.791 1.013,-0.693 1.771,-3.42 2.995,-1.995 0.496,0.578 1.042,1.114 1.483,1.737 0.559,0.789 1.71,0.83 2.292,1.626 1.109,-2.146 -1.817,-3.451 -2.804,-4.625 -0.44,-0.523 -1.126,-1.101 -1.644,-1.519 -0.31,-0.251 -2.023,-1.063 -2.393,-1.055 0.265,0.687 0.048,2.505 0.349,2.909 0.744,0.998 -0.136,1.049 -0.208,2.131"
-              id="path5872"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2239.488,696.456 c 0.27,0.527 0.63,0.983 1.08,1.368 -0.145,-0.972 -0.415,-2.162 -1.37,-2.664 -0.218,0.503 -0.121,0.935 0.29,1.296"
-              id="path5876"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2239.638,718.128 c 0.104,1.506 -0.364,2.977 -0.51,4.464 2.061,0.544 1.624,-4.352 2.455,-5.322 1.029,-1.203 2.119,-2.178 2.298,-3.852 0.213,-1.993 -0.422,-3.956 -0.362,-5.946 -0.997,0.205 -1.279,3.326 -1.448,4.085 -0.515,2.302 -1.959,4.264 -2.433,6.571"
-              id="path5878"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2241.508,690.264 c 0.119,-0.492 -0.121,-0.589 -0.72,-0.288 0.24,0.095 0.48,0.192 0.72,0.288"
-              id="path5880"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2242.588,687.672 c 1.809,-1.271 1.543,-4.132 -0.87,-4.608 -0.706,1.478 -0.232,3.594 0.87,4.608"
-              id="path5884"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2243.019,690.696 c -0.229,-0.193 -0.493,-0.241 -0.79,-0.145 0.263,0.049 0.526,0.097 0.79,0.145"
-              id="path5886"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2243.229,720.576 c -0.163,2.803 4.839,-0.383 5.479,-0.145 -0.395,-1.444 -0.448,-4.291 -3.035,-3.583 -0.908,0.248 -3.694,2.932 -2.444,3.728"
-              id="path5890"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2244.168,696.456 c 0.568,1.532 1.226,2.841 2.521,3.888 0.722,0.583 2.927,2.916 3.819,2.521 -0.304,0.713 -0.27,1.659 -1.16,1.943 0.442,0.363 0.803,0.795 1.08,1.296 -0.767,-0.844 -1.213,0.229 -1.839,0.593 -0.97,0.563 -2.492,-1 -2.911,-1.744 -0.721,0.887 0.645,5.152 1.73,5.615 -0.191,-1.934 2.081,0.562 2.133,1.548 0.133,2.513 -0.325,4.446 2.337,5.652 -0.114,-0.759 -0.505,-1.451 -0.51,-2.231 0.421,0.469 1.888,0.508 2.029,-0.29 0.078,-0.443 -1,-2.532 -1.268,-2.979 -0.614,-1.03 -0.801,-0.895 -0.901,-2.274 -0.065,-0.893 0.13,-2.297 -0.65,-2.952 0.598,-0.155 2.289,1.494 2.305,1.476 0.66,-0.774 3.139,0.212 3.815,0.612 -0.546,-0.639 -0.923,-1.537 -1.8,-1.8 1.133,-0.867 -1.138,-3.423 -1.241,-4.453 -0.11,-1.085 0.378,-2.05 -0.217,-2.974 -0.38,-0.59 1.05,-0.418 -0.099,-1.109 -0.623,-0.375 -1.418,-0.098 -1.36,-1.081 0.073,-1.262 -1.022,-1.306 -1.925,-1.504 -0.977,-0.214 -1.291,-0.23 -2.32,0.101 -1.183,0.378 -2.482,-0.673 -3.568,0.146"
-              id="path5894"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2246.688,705.456 c 1.272,0.186 0.341,-2.777 -1.221,-1.872 0.34,0.653 0.527,1.476 1.221,1.872"
-              id="path5896"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2253.238,718.128 c 0.271,0.234 0.538,0.474 0.8,0.72 -0.147,-0.683 -0.478,-1.826 -1.44,-1.512 0.188,0.284 0.402,0.548 0.64,0.792"
-              id="path5904"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2256.118,718.128 c 0.575,0.399 0.647,1.009 0.15,1.512 1.073,0.158 0.827,-1.373 0.7,-1.963 -0.134,-0.618 -0.527,-1.164 -0.631,-1.781 -0.095,-0.562 0.283,-1.014 0.07,-1.584 -1.419,0.606 -0.285,1.812 -1.08,2.736 0.574,0.093 0.956,0.471 0.791,1.08"
-              id="path5906"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2259.218,721.944 c 0.494,-0.511 0.149,-1.194 -0.359,-1.513 -0.084,0.553 0.036,1.057 0.359,1.513"
-              id="path5912"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2259.648,718.128 c -0.359,0.484 -0.672,2.522 0.475,2.26 1.164,-0.266 0.301,-2.116 -0.475,-2.26"
-              id="path5914"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2246.118,713.088 c -0.05,-0.264 -0.101,-0.528 -0.15,-0.792 0.817,0.12 0.867,0.384 0.15,0.792"
-              id="path5920"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2250.428,706.104 c 0.127,0.3 0.177,0.611 0.15,0.937 -0.145,-0.297 -0.196,-0.609 -0.15,-0.937"
-              id="path5924"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2240.278,709.2 c -0.529,-2.814 3.298,-1.367 0,0"
-              id="path5936"
-              data-inkscape-connector-curvature="0"
-            />
-          </g>
-          <g
-            transform="matrix(6.4876122,0,0,6.4876122,-12925.057,-3966.0743)"
-            style={{
-              fill: clickedCountryId === "Cambodia" ? "#FF0000" : "#005500",
-              fillOpacity: 0.98823529,
-              stroke: "#ffffff",
-              strokeWidth: 0.5257113,
-              strokeMiterlimit: 3.98288536,
-              strokeDasharray: "none",
-              strokeOpacity: 1,
-            }}
-            id="Cambodia"
-            className="country"
-          >
-            <path
-              d="m 2091.818,714.312 c -0.314,0.455 -0.748,0.695 -1.301,0.721 0.975,0.748 1.068,2.032 1.801,2.951 -0.359,-1.23 0.682,-2.687 -0.5,-3.672"
-              id="path3472"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2083.678,705.96 c -0.144,0 -0.286,0 -0.43,0 0.157,0.291 0.351,0.555 0.58,0.792 -0.05,-0.264 -0.1,-0.528 -0.15,-0.792"
-              id="path6024"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2085.988,712.296 c 0.351,-0.609 0.134,-0.849 -0.65,-0.72 0.2,0.256 0.417,0.495 0.65,0.72"
-              id="path6026"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2077.059,689.544 c 0.021,-0.754 -0.021,-1.519 -0.291,-2.232 2.388,0.082 3.398,-3.35 4.134,-5.084 0.803,-1.893 4.694,-3.048 6.358,-2 0.818,0.515 2.127,0.265 2.994,-0.038 1.146,-0.401 2.048,0.271 3.246,-0.219 2.161,-0.883 3.406,2.429 4.368,0.502 0.146,3.016 3.734,0.908 4.929,2.909 0.585,0.98 2.597,1.03 2.407,-0.463 -0.105,-0.824 -1.93,-1.834 -0.832,-2.604 0.46,-0.323 0.907,0.328 1.051,-0.422 0.118,-0.612 0.634,-0.398 1.099,-0.438 0.584,-0.05 0.502,-0.866 0.988,-0.909 0.66,-0.057 1.327,0.861 1.799,1.206 2.073,1.516 4.866,-0.699 5.909,-2.376 0.01,0.553 0.126,1.191 -0.069,1.728 -0.215,0.59 -1.113,0.875 -0.884,1.627 0.218,0.715 0.026,1.459 0.218,2.179 0.201,0.754 0.743,1.65 1.041,2.396 0.723,1.804 1.818,2.505 1.283,4.661 -0.245,0.984 -0.771,1.873 -0.406,2.889 0.282,0.787 0.848,1.536 0.93,2.386 0.079,0.816 -0.295,1.968 -0.625,2.698 -0.448,0.989 -0.682,-0.292 -1.436,0.115 -0.946,0.512 -1.562,1.457 -2.457,2.029 -1.291,0.824 -2.323,0.688 -3.784,1.053 0.635,0.769 0.043,1.756 0.5,2.592 -0.818,-0.159 -2.956,-1.61 -3.598,-0.614 -0.421,0.653 -1.31,0.751 -1.129,1.597 0.189,0.892 0.502,1.904 0.968,2.692 0.544,0.922 2.206,0.972 2.389,2.302 -0.943,0.138 -0.12,0.796 -0.141,1.584 -0.668,-0.215 -1.451,-0.193 -2.057,-0.59 -0.525,-0.344 -0.767,-1.402 -1.54,-1.251 -1.475,0.29 -3.387,1.714 -4.753,0.4 -0.763,0.79 0.175,1.853 -0.811,2.834 -0.957,0.953 -1.826,0.476 -2.934,0.92 -0.818,0.328 -0.936,0.854 -1.945,0.423 -0.993,-0.423 -3.031,-0.719 -3.53,-1.729 0.271,1.068 -1.2,1.8 -1.974,1.104 -1.023,-0.922 -0.089,-1.119 0.509,-1.752 0.698,-0.741 -0.152,-2.717 -0.799,-3.234 -1.156,-0.926 -1.288,1.493 -1.477,2.083 -0.711,-0.599 -1.151,0.273 -1.806,-0.048 -0.991,-0.487 -0.391,-2.059 -0.629,-2.875 -0.217,-0.742 0.512,-1.807 -0.259,-2.347 -0.537,-0.376 -1.686,-1.684 -0.337,-1.858 -0.57,-0.302 -0.959,-0.009 -1.08,0.576 0.186,-1.228 -0.867,-2.321 -1.289,-3.385 -0.471,-1.19 0.146,-2.854 -0.591,-3.917 -1.575,-2.27 -2.75,-4.422 -3.657,-7.1"
-              id="path6044"
-              data-inkscape-connector-curvature="0"
-            />
-          </g>
-          <g
-            //transform="matrix(6.4876122,0,0,6.4876122,-12925.057,-3966.0743)"
-            style={{
-              fill: clickedCountryId === "Laos" ? "#FF0000" : "#005500",
-              fillOpacity: 0.98823529,
-              stroke: "#ffffff",
-              strokeWidth: 0.7,
-              strokeMiterlimit: 3.98288536,
-              strokeDasharray: "none",
-              strokeOpacity: 1,
-            }}
-            id="Laos"
-            className="country"
-          >
-            <path
-              d="m 392.18039,111.77972 c 1.86196,-7.5711 -0.55142,-18.411898 6.55896,-23.822598 4.07418,-3.101 15.14862,0.3958 15.88811,-6.5395 -9.60798,5.6962 -1.93964,-11.4247 -0.72641,-14.5322 2.00451,-5.1382 1.94618,-7.9344 7.53206,-10.4126 4.70355,-2.089 13.7732,-6.8769 12.33291,-13.358 8.01234,4.5024 1.27156,26.0218 15.41465,20.2219 8.27155,-3.3931 8.47925,0.4801 15.67413,3.2438 10.15287,3.899 1.02489,-8.4599 0.78502,-9.2579 -1.97889,-6.4876 1.79054,-11.1976 0.96655,-17.1856 -0.83054,-5.9946 -5.74151,-11.6712 -9.27081,-16.3099 -3.97687,-5.2355 -7.49962,-10.2699 -7.92784,-17.1207996 -0.0981,-1.5441 4.88509,-18.4702004 7.66835,-15.0513 2.80261,3.4385 2.1994,5.4951 8.41438,3.8926 9.58863,-2.4783 10.60724,4.4246 16.10889,9.926 5.8192,5.8258996 12.51445,11.3014996 18.66483,16.8353996 4.71636,4.2364 4.1584,13.7278 10.56827,15.6676 2.68595,-2.874 3.31505,-6.1438 1.88131,-9.8093 1.49861,5.8064 8.18105,4.5738 9.59517,11.1198 1.42093,6.52 -2.6208,14.156 -0.66181,21.6946 5.73524,22.0254 28.89592,23.018 45.23814,33.0479 4.05482,-5.4756 2.56274,-10.5878 9.79633,-13.5462 8.64798,-3.5357 15.06439,-0.2205 22.11644,4.8592 4.08699,2.9454 10.09453,6.2346 10.92507,11.7491 0.73949,4.865698 -7.92784,4.755398 -9.16696,9.549798 4.32712,1.7646 12.346,0.4736 14.03269,6.1502 1.29091,4.3467 4.55418,4.191 7.8954,5.9946 14.12346,7.5905 8.93336,8.9529 1.88131,19.6185 9.60172,2.7508 -8.95272,14.1495 -15.80361,13.1115 -1.66734,-0.253 -18.51573,-5.4042 -19.68344,-1.9009 -1.129,3.3606 5.98792,9.1216 4.67084,14.954 -5.49483,0.052 -10.96404,6.1567 -3.20466,8.5506 12.98164,4.0159 26.95601,16.21908 38.39351,24.22479 4.50893,3.15947 26.13856,3.58114 25.5547,11.41169 -0.45411,6.14378 -4.45687,7.92138 0.90849,13.78618 3.94443,4.32074 10.6266,11.07436 15.36259,11.18464 4.86573,0.11039 7.90848,8.60908 9.58863,12.07994 3.84058,7.94732 9.61453,13.8251 15.6417,20.11808 3.46442,3.6201 16.00477,20.76686 21.27936,19.20337 5.13177,-1.51162 9.25773,15.3302 14.31808,17.49709 5.97511,2.5561 3.59417,14.85014 7.40231,19.83262 1.25875,1.64785 4.28842,8.01868 6.5527,8.37551 2.74401,0.43467 5.69599,-2.48476 8.38848,-0.92125 5.45586,3.17244 3.62007,8.42092 10.64595,10.06229 6.62384,1.53756 13.83181,11.44416 21.14335,14.22733 -2.45235,5.57936 -9.90673,8.10951 -12.02163,13.88349 -2.29672,6.27352 14.85014,14.32465 17.33493,18.88545 3.77571,6.95471 11.89162,6.92876 7.83707,17.02349 -1.20669,2.99728 -4.56099,6.26055 -4.6583,9.64058 -0.0845,3.15298 3.41918,5.32633 3.3802,9.05671 -0.13084,10.58778 -2.62108,12.23563 -11.17829,14.26626 -7.35679,1.74519 -9.30952,11.48308 -19.42395,11.63229 -9.75081,0.14937 -11.63867,-12.7287 -17.72417,-13.57207 -4.0352,-0.55793 -2.00451,4.95653 -4.67738,6.07888 -2.66006,1.11587 -5.71562,1.3689 -8.49234,2.10199 -3.25018,0.84989 -12.65728,5.57287 -10.43198,9.65356 2.95825,5.4042 10.20493,11.51552 4.82021,17.6593 -6.58486,7.50616 -21.96053,-13.43587 -32.75585,-9.47841 -6.07242,2.22525 -15.68722,-8.58959 -9.79006,-13.54614 4.41162,-3.69794 12.5537,-7.13636 11.82047,-14.02621 -0.67462,-6.40327 1.82298,-15.8687 0.44784,-21.31831 -0.9213,-3.62655 -7.0068,-4.05474 -7.0068,-8.17438 0,-3.78876 3.23737,-7.22721 4.67765,-10.51642 -1.82952,0.88232 -3.45134,0.77852 -4.87227,-0.30491 -0.94066,-5.24199 11.21727,-12.0994 3.27635,-20.15052 -4.30123,-4.3597 -21.34424,-9.52381 -10.53583,-16.90672 -18.00983,-7.06503 -36.43452,-19.1774 -39.2892,-40.52164 -2.21875,-16.57584 3.43199,-39.79501 -13.60447,-49.89623 -17.36737,-10.30234 -23.27107,-35.30559 -43.42797,-39.60688 -2.60145,-0.55793 -37.15466,-10.69807 -25.58059,2.0436 -8.129,1.13533 -5.72216,12.70275 -9.34223,16.34879 -1.42093,1.44023 -9.56274,1.59595 -12.80665,4.51536 -5.02137,4.52837 -6.96101,-0.3568 -10.53583,-3.01023 -5.96203,-4.41808 -15.71311,-10.7954 -22.80387,-13.02065 -7.21424,-2.26417 -7.47373,12.43025 -16.01159,7.31155 -0.62283,12.83897 -16.03093,21.77891 -23.47849,30.67343 -3.43853,4.10665 -8.31081,1.83599 -12.04099,-2.05009 -4.14559,-4.32723 -0.83054,-12.30051 1.03796,-16.84834 2.91301,-7.09094 3.8016,-11.8788 2.70531,-19.57312 -0.65527,-4.59324 -5.18356,-12.48218 -4.61933,-16.49151 1.16799,-2.32255 2.64071,-4.44403 4.42471,-6.35137 2.3744,-3.97692 -0.0845,-4.10018 0.18808,-8.13546 0.59694,-8.83613 3.52929,-13.94189 -0.92758,-22.3953 -6.28012,-11.9242 0.19353,-30.0506 -19.96255,-24.5101 -2.40683,-0.014 -4.19738,0.9342 -5.39126,2.861 -1.96553,4.5283 -2.52349,0.6033 -6.35126,1.3624 -6.69525,1.3364 -9.77044,-1.3559 -13.0596,-7.7981 -4.68392,-9.167 7.60347,-20.2738 2.23839,-27.0793 -2.20595,-2.7962 -8.62863,-11.6647 -11.6324,-13.0661 -5.9882,-2.7896 -7.11039,4.2105 -12.13176,5.6053 -0.31809,-1.5635 -0.64246,-3.1205 -0.96682,-4.6775"
-              style={{
-                fill: "#005500",
-                fillOpacity: 0.98823529,
-                stroke: "#ffffff",
-                strokeWidth: 0.7,
-                strokeMiterlimit: 3.98288536,
-                strokeDasharray: "none",
-                strokeOpacity: 1,
-              }}
-              id="path5591"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 392.18039,111.77972 c 1.86196,-7.5711 -0.55142,-18.411898 6.55896,-23.822598 4.07418,-3.101 15.14862,0.3958 15.88811,-6.5395 -9.60798,5.6962 -1.93964,-11.4247 -0.72641,-14.5322 2.00451,-5.1382 1.94618,-7.9344 7.53206,-10.4126 4.70355,-2.089 13.7732,-6.8769 12.33291,-13.358 8.01234,4.5024 1.27156,26.0218 15.41465,20.2219 8.27155,-3.3931 8.47925,0.4801 15.67413,3.2438 10.15287,3.899 1.02489,-8.4599 0.78502,-9.2579 -1.97889,-6.4876 1.79054,-11.1976 0.96655,-17.1856 -0.83054,-5.9946 -5.74151,-11.6712 -9.27081,-16.3099 -3.97687,-5.2355 -7.49962,-10.2699 -7.92784,-17.1207996 -0.0981,-1.5441 4.88509,-18.4702004 7.66835,-15.0513 2.80261,3.4385 2.1994,5.4951 8.41438,3.8926 9.58863,-2.4783 10.60724,4.4246 16.10889,9.926 5.8192,5.8258996 12.51445,11.3014996 18.66483,16.8353996 4.71636,4.2364 4.1584,13.7278 10.56827,15.6676 2.68595,-2.874 3.31505,-6.1438 1.88131,-9.8093 1.49861,5.8064 8.18105,4.5738 9.59517,11.1198 1.42093,6.52 -2.6208,14.156 -0.66181,21.6946 5.73524,22.0254 28.89592,23.018 45.23814,33.0479 4.05482,-5.4756 2.56274,-10.5878 9.79633,-13.5462 8.64798,-3.5357 15.06439,-0.2205 22.11644,4.8592 4.08699,2.9454 10.09453,6.2346 10.92507,11.7491 0.73949,4.865698 -7.92784,4.755398 -9.16696,9.549798 4.32712,1.7646 12.346,0.4736 14.03269,6.1502 1.29091,4.3467 4.55418,4.191 7.8954,5.9946 14.12346,7.5905 8.93336,8.9529 1.88131,19.6185 9.60172,2.7508 -8.95272,14.1495 -15.80361,13.1115 -1.66734,-0.253 -18.51573,-5.4042 -19.68344,-1.9009 -1.129,3.3606 5.98792,9.1216 4.67084,14.954 -5.49483,0.052 -10.96404,6.1567 -3.20466,8.5506 12.98164,4.0159 26.95601,16.21908 38.39351,24.22479 4.50893,3.15947 26.13856,3.58114 25.5547,11.41169 -0.45411,6.14378 -4.45687,7.92138 0.90849,13.78618 3.94443,4.32074 10.6266,11.07436 15.36259,11.18464 4.86573,0.11039 7.90848,8.60908 9.58863,12.07994 3.84058,7.94732 9.61453,13.8251 15.6417,20.11808 3.46442,3.6201 16.00477,20.76686 21.27936,19.20337 5.13177,-1.51162 9.25773,15.3302 14.31808,17.49709 5.97511,2.5561 3.59417,14.85014 7.40231,19.83262 1.25875,1.64785 4.28842,8.01868 6.5527,8.37551 2.74401,0.43467 5.69599,-2.48476 8.38848,-0.92125 5.45586,3.17244 3.62007,8.42092 10.64595,10.06229 6.62384,1.53756 13.83181,11.44416 21.14335,14.22733 -2.45235,5.57936 -9.90673,8.10951 -12.02163,13.88349 -2.29672,6.27352 14.85014,14.32465 17.33493,18.88545 3.77571,6.95471 11.89162,6.92876 7.83707,17.02349 -1.20669,2.99728 -4.56099,6.26055 -4.6583,9.64058 -0.0845,3.15298 3.41918,5.32633 3.3802,9.05671 -0.13084,10.58778 -2.62108,12.23563 -11.17829,14.26626 -7.35679,1.74519 -9.30952,11.48308 -19.42395,11.63229 -9.75081,0.14937 -11.63867,-12.7287 -17.72417,-13.57207 -4.0352,-0.55793 -2.00451,4.95653 -4.67738,6.07888 -2.66006,1.11587 -5.71562,1.3689 -8.49234,2.10199 -3.25018,0.84989 -12.65728,5.57287 -10.43198,9.65356 2.95825,5.4042 10.20493,11.51552 4.82021,17.6593 -6.58486,7.50616 -21.96053,-13.43587 -32.75585,-9.47841 -6.07242,2.22525 -15.68722,-8.58959 -9.79006,-13.54614 4.41162,-3.69794 12.5537,-7.13636 11.82047,-14.02621 -0.67462,-6.40327 1.82298,-15.8687 0.44784,-21.31831 -0.9213,-3.62655 -7.0068,-4.05474 -7.0068,-8.17438 0,-3.78876 3.23737,-7.22721 4.67765,-10.51642 -1.82952,0.88232 -3.45134,0.77852 -4.87227,-0.30491 -0.94066,-5.24199 11.21727,-12.0994 3.27635,-20.15052 -4.30123,-4.3597 -21.34424,-9.52381 -10.53583,-16.90672 -18.00983,-7.06503 -36.43452,-19.1774 -39.2892,-40.52164 -2.21875,-16.57584 3.43199,-39.79501 -13.60447,-49.89623 -17.36737,-10.30234 -23.27107,-35.30559 -43.42797,-39.60688 -2.60145,-0.55793 -37.15466,-10.69807 -25.58059,2.0436 -8.129,1.13533 -5.72216,12.70275 -9.34223,16.34879 -1.42093,1.44023 -9.56274,1.59595 -12.80665,4.51536 -5.02137,4.52837 -6.96101,-0.3568 -10.53583,-3.01023 -5.96203,-4.41808 -15.71311,-10.7954 -22.80387,-13.02065 -7.21424,-2.26417 -7.47373,12.43025 -16.01159,7.31155 -0.62283,12.83897 -16.03093,21.77891 -23.47849,30.67343 -3.43853,4.10665 -8.31081,1.83599 -12.04099,-2.05009 -4.14559,-4.32723 -0.83054,-12.30051 1.03796,-16.84834 2.91301,-7.09094 3.8016,-11.8788 2.70531,-19.57312 -0.65527,-4.59324 -5.18356,-12.48218 -4.61933,-16.49151 1.16799,-2.32255 2.64071,-4.44403 4.42471,-6.35137 2.3744,-3.97692 -0.0845,-4.10018 0.18808,-8.13546 0.59694,-8.83613 3.52929,-13.94189 -0.92758,-22.3953 -6.28012,-11.9242 0.19353,-30.0506 -19.96255,-24.5101 -2.40683,-0.014 -4.19738,0.9342 -5.39126,2.861 -1.96553,4.5283 -2.52349,0.6033 -6.35126,1.3624 -6.69525,1.3364 -9.77044,-1.3559 -13.0596,-7.7981 -4.68392,-9.167 7.60347,-20.2738 2.23839,-27.0793 -2.20595,-2.7962 -8.62863,-11.6647 -11.6324,-13.0661 -5.9882,-2.7896 -7.11039,4.2105 -12.13176,5.6053 -0.31809,-1.5635 -0.64246,-3.1205 -0.96682,-4.6775"
-              style={{
-                fill: "#005500",
-                fillOpacity: 0.98823529,
-                stroke: "#ffffff",
-                strokeWidth: 0.7,
-                strokeMiterlimit: 3.98288536,
-                strokeDasharray: "none",
-                strokeOpacity: 1,
-              }}
-              id="path5593"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 392.18039,111.77972 c 1.86196,-7.5711 -0.55142,-18.411898 6.55896,-23.822598 4.07418,-3.101 15.14862,0.3958 15.88811,-6.5395 -9.60798,5.6962 -1.93964,-11.4247 -0.72641,-14.5322 2.00451,-5.1382 1.94618,-7.9344 7.53206,-10.4126 4.70355,-2.089 13.7732,-6.8769 12.33291,-13.358 8.01234,4.5024 1.27156,26.0218 15.41465,20.2219 8.27155,-3.3931 8.47925,0.4801 15.67413,3.2438 10.15287,3.899 1.02489,-8.4599 0.78502,-9.2579 -1.97889,-6.4876 1.79054,-11.1976 0.96655,-17.1856 -0.83054,-5.9946 -5.74151,-11.6712 -9.27081,-16.3099 -3.97687,-5.2355 -7.49962,-10.2699 -7.92784,-17.1207996 -0.0981,-1.5441 4.88509,-18.4702004 7.66835,-15.0513 2.80261,3.4385 2.1994,5.4951 8.41438,3.8926 9.58863,-2.4783 10.60724,4.4246 16.10889,9.926 5.8192,5.8258996 12.51445,11.3014996 18.66483,16.8353996 4.71636,4.2364 4.1584,13.7278 10.56827,15.6676 2.68595,-2.874 3.31505,-6.1438 1.88131,-9.8093 1.49861,5.8064 8.18105,4.5738 9.59517,11.1198 1.42093,6.52 -2.6208,14.156 -0.66181,21.6946 5.73524,22.0254 28.89592,23.018 45.23814,33.0479 4.05482,-5.4756 2.56274,-10.5878 9.79633,-13.5462 8.64798,-3.5357 15.06439,-0.2205 22.11644,4.8592 4.08699,2.9454 10.09453,6.2346 10.92507,11.7491 0.73949,4.865698 -7.92784,4.755398 -9.16696,9.549798 4.32712,1.7646 12.346,0.4736 14.03269,6.1502 1.29091,4.3467 4.55418,4.191 7.8954,5.9946 14.12346,7.5905 8.93336,8.9529 1.88131,19.6185 9.60172,2.7508 -8.95272,14.1495 -15.80361,13.1115 -1.66734,-0.253 -18.51573,-5.4042 -19.68344,-1.9009 -1.129,3.3606 5.98792,9.1216 4.67084,14.954 -5.49483,0.052 -10.96404,6.1567 -3.20466,8.5506 12.98164,4.0159 26.95601,16.21908 38.39351,24.22479 4.50893,3.15947 26.13856,3.58114 25.5547,11.41169 -0.45411,6.14378 -4.45687,7.92138 0.90849,13.78618 3.94443,4.32074 10.6266,11.07436 15.36259,11.18464 4.86573,0.11039 7.90848,8.60908 9.58863,12.07994 3.84058,7.94732 9.61453,13.8251 15.6417,20.11808 3.46442,3.6201 16.00477,20.76686 21.27936,19.20337 5.13177,-1.51162 9.25773,15.3302 14.31808,17.49709 5.97511,2.5561 3.59417,14.85014 7.40231,19.83262 1.25875,1.64785 4.28842,8.01868 6.5527,8.37551 2.74401,0.43467 5.69599,-2.48476 8.38848,-0.92125 5.45586,3.17244 3.62007,8.42092 10.64595,10.06229 6.62384,1.53756 13.83181,11.44416 21.14335,14.22733 -2.45235,5.57936 -9.90673,8.10951 -12.02163,13.88349 -2.29672,6.27352 14.85014,14.32465 17.33493,18.88545 3.77571,6.95471 11.89162,6.92876 7.83707,17.02349 -1.20669,2.99728 -4.56099,6.26055 -4.6583,9.64058 -0.0845,3.15298 3.41918,5.32633 3.3802,9.05671 -0.13084,10.58778 -2.62108,12.23563 -11.17829,14.26626 -7.35679,1.74519 -9.30952,11.48308 -19.42395,11.63229 -9.75081,0.14937 -11.63867,-12.7287 -17.72417,-13.57207 -4.0352,-0.55793 -2.00451,4.95653 -4.67738,6.07888 -2.66006,1.11587 -5.71562,1.3689 -8.49234,2.10199 -3.25018,0.84989 -12.65728,5.57287 -10.43198,9.65356 2.95825,5.4042 10.20493,11.51552 4.82021,17.6593 -6.58486,7.50616 -21.96053,-13.43587 -32.75585,-9.47841 -6.07242,2.22525 -15.68722,-8.58959 -9.79006,-13.54614 4.41162,-3.69794 12.5537,-7.13636 11.82047,-14.02621 -0.67462,-6.40327 1.82298,-15.8687 0.44784,-21.31831 -0.9213,-3.62655 -7.0068,-4.05474 -7.0068,-8.17438 0,-3.78876 3.23737,-7.22721 4.67765,-10.51642 -1.82952,0.88232 -3.45134,0.77852 -4.87227,-0.30491 -0.94066,-5.24199 11.21727,-12.0994 3.27635,-20.15052 -4.30123,-4.3597 -21.34424,-9.52381 -10.53583,-16.90672 -18.00983,-7.06503 -36.43452,-19.1774 -39.2892,-40.52164 -2.21875,-16.57584 3.43199,-39.79501 -13.60447,-49.89623 -17.36737,-10.30234 -23.27107,-35.30559 -43.42797,-39.60688 -2.60145,-0.55793 -37.15466,-10.69807 -25.58059,2.0436 -8.129,1.13533 -5.72216,12.70275 -9.34223,16.34879 -1.42093,1.44023 -9.56274,1.59595 -12.80665,4.51536 -5.02137,4.52837 -6.96101,-0.3568 -10.53583,-3.01023 -5.96203,-4.41808 -15.71311,-10.7954 -22.80387,-13.02065 -7.21424,-2.26417 -7.47373,12.43025 -16.01159,7.31155 -0.62283,12.83897 -16.03093,21.77891 -23.47849,30.67343 -3.43853,4.10665 -8.31081,1.83599 -12.04099,-2.05009 -4.14559,-4.32723 -0.83054,-12.30051 1.03796,-16.84834 2.91301,-7.09094 3.8016,-11.8788 2.70531,-19.57312 -0.65527,-4.59324 -5.18356,-12.48218 -4.61933,-16.49151 1.16799,-2.32255 2.64071,-4.44403 4.42471,-6.35137 2.3744,-3.97692 -0.0845,-4.10018 0.18808,-8.13546 0.59694,-8.83613 3.52929,-13.94189 -0.92758,-22.3953 -6.28012,-11.9242 0.19353,-30.0506 -19.96255,-24.5101 -2.40683,-0.014 -4.19738,0.9342 -5.39126,2.861 -1.96553,4.5283 -2.52349,0.6033 -6.35126,1.3624 -6.69525,1.3364 -9.77044,-1.3559 -13.0596,-7.7981 -4.68392,-9.167 7.60347,-20.2738 2.23839,-27.0793 -2.20595,-2.7962 -8.62863,-11.6647 -11.6324,-13.0661 -5.9882,-2.7896 -7.11039,4.2105 -12.13176,5.6053 -0.31809,-1.5635 -0.64246,-3.1205 -0.96682,-4.6775"
-              style={{
-                fill: "#005500",
-                fillOpacity: 0.98823529,
-                stroke: "#ffffff",
-                strokeWidth: 0.7,
-                strokeMiterlimit: 3.98288536,
-                strokeDasharray: "none",
-                strokeOpacity: 1,
-              }}
-              id="path5595"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 392.18039,111.77972 c 1.86196,-7.5711 -0.55142,-18.411898 6.55896,-23.822598 4.07418,-3.101 15.14862,0.3958 15.88811,-6.5395 -9.60798,5.6962 -1.93964,-11.4247 -0.72641,-14.5322 2.00451,-5.1382 1.94618,-7.9344 7.53206,-10.4126 4.70355,-2.089 13.7732,-6.8769 12.33291,-13.358 8.01234,4.5024 1.27156,26.0218 15.41465,20.2219 8.27155,-3.3931 8.47925,0.4801 15.67413,3.2438 10.15287,3.899 1.02489,-8.4599 0.78502,-9.2579 -1.97889,-6.4876 1.79054,-11.1976 0.96655,-17.1856 -0.83054,-5.9946 -5.74151,-11.6712 -9.27081,-16.3099 -3.97687,-5.2355 -7.49962,-10.2699 -7.92784,-17.1207996 -0.0981,-1.5441 4.88509,-18.4702004 7.66835,-15.0513 2.80261,3.4385 2.1994,5.4951 8.41438,3.8926 9.58863,-2.4783 10.60724,4.4246 16.10889,9.926 5.8192,5.8258996 12.51445,11.3014996 18.66483,16.8353996 4.71636,4.2364 4.1584,13.7278 10.56827,15.6676 2.68595,-2.874 3.31505,-6.1438 1.88131,-9.8093 1.49861,5.8064 8.18105,4.5738 9.59517,11.1198 1.42093,6.52 -2.6208,14.156 -0.66181,21.6946 5.73524,22.0254 28.89592,23.018 45.23814,33.0479 4.05482,-5.4756 2.56274,-10.5878 9.79633,-13.5462 8.64798,-3.5357 15.06439,-0.2205 22.11644,4.8592 4.08699,2.9454 10.09453,6.2346 10.92507,11.7491 0.73949,4.865698 -7.92784,4.755398 -9.16696,9.549798 4.32712,1.7646 12.346,0.4736 14.03269,6.1502 1.29091,4.3467 4.55418,4.191 7.8954,5.9946 14.12346,7.5905 8.93336,8.9529 1.88131,19.6185 9.60172,2.7508 -8.95272,14.1495 -15.80361,13.1115 -1.66734,-0.253 -18.51573,-5.4042 -19.68344,-1.9009 -1.129,3.3606 5.98792,9.1216 4.67084,14.954 -5.49483,0.052 -10.96404,6.1567 -3.20466,8.5506 12.98164,4.0159 26.95601,16.21908 38.39351,24.22479 4.50893,3.15947 26.13856,3.58114 25.5547,11.41169 -0.45411,6.14378 -4.45687,7.92138 0.90849,13.78618 3.94443,4.32074 10.6266,11.07436 15.36259,11.18464 4.86573,0.11039 7.90848,8.60908 9.58863,12.07994 3.84058,7.94732 9.61453,13.8251 15.6417,20.11808 3.46442,3.6201 16.00477,20.76686 21.27936,19.20337 5.13177,-1.51162 9.25773,15.3302 14.31808,17.49709 5.97511,2.5561 3.59417,14.85014 7.40231,19.83262 1.25875,1.64785 4.28842,8.01868 6.5527,8.37551 2.74401,0.43467 5.69599,-2.48476 8.38848,-0.92125 5.45586,3.17244 3.62007,8.42092 10.64595,10.06229 6.62384,1.53756 13.83181,11.44416 21.14335,14.22733 -2.45235,5.57936 -9.90673,8.10951 -12.02163,13.88349 -2.29672,6.27352 14.85014,14.32465 17.33493,18.88545 3.77571,6.95471 11.89162,6.92876 7.83707,17.02349 -1.20669,2.99728 -4.56099,6.26055 -4.6583,9.64058 -0.0845,3.15298 3.41918,5.32633 3.3802,9.05671 -0.13084,10.58778 -2.62108,12.23563 -11.17829,14.26626 -7.35679,1.74519 -9.30952,11.48308 -19.42395,11.63229 -9.75081,0.14937 -11.63867,-12.7287 -17.72417,-13.57207 -4.0352,-0.55793 -2.00451,4.95653 -4.67738,6.07888 -2.66006,1.11587 -5.71562,1.3689 -8.49234,2.10199 -3.25018,0.84989 -12.65728,5.57287 -10.43198,9.65356 2.95825,5.4042 10.20493,11.51552 4.82021,17.6593 -6.58486,7.50616 -21.96053,-13.43587 -32.75585,-9.47841 -6.07242,2.22525 -15.68722,-8.58959 -9.79006,-13.54614 4.41162,-3.69794 12.5537,-7.13636 11.82047,-14.02621 -0.67462,-6.40327 1.82298,-15.8687 0.44784,-21.31831 -0.9213,-3.62655 -7.0068,-4.05474 -7.0068,-8.17438 0,-3.78876 3.23737,-7.22721 4.67765,-10.51642 -1.82952,0.88232 -3.45134,0.77852 -4.87227,-0.30491 -0.94066,-5.24199 11.21727,-12.0994 3.27635,-20.15052 -4.30123,-4.3597 -21.34424,-9.52381 -10.53583,-16.90672 -18.00983,-7.06503 -36.43452,-19.1774 -39.2892,-40.52164 -2.21875,-16.57584 3.43199,-39.79501 -13.60447,-49.89623 -17.36737,-10.30234 -23.27107,-35.30559 -43.42797,-39.60688 -2.60145,-0.55793 -37.15466,-10.69807 -25.58059,2.0436 -8.129,1.13533 -5.72216,12.70275 -9.34223,16.34879 -1.42093,1.44023 -9.56274,1.59595 -12.80665,4.51536 -5.02137,4.52837 -6.96101,-0.3568 -10.53583,-3.01023 -5.96203,-4.41808 -15.71311,-10.7954 -22.80387,-13.02065 -7.21424,-2.26417 -7.47373,12.43025 -16.01159,7.31155 -0.62283,12.83897 -16.03093,21.77891 -23.47849,30.67343 -3.43853,4.10665 -8.31081,1.83599 -12.04099,-2.05009 -4.14559,-4.32723 -0.83054,-12.30051 1.03796,-16.84834 2.91301,-7.09094 3.8016,-11.8788 2.70531,-19.57312 -0.65527,-4.59324 -5.18356,-12.48218 -4.61933,-16.49151 1.16799,-2.32255 2.64071,-4.44403 4.42471,-6.35137 2.3744,-3.97692 -0.0845,-4.10018 0.18808,-8.13546 0.59694,-8.83613 3.52929,-13.94189 -0.92758,-22.3953 -6.28012,-11.9242 0.19353,-30.0506 -19.96255,-24.5101 -2.40683,-0.014 -4.19738,0.9342 -5.39126,2.861 -1.96553,4.5283 -2.52349,0.6033 -6.35126,1.3624 -6.69525,1.3364 -9.77044,-1.3559 -13.0596,-7.7981 -4.68392,-9.167 7.60347,-20.2738 2.23839,-27.0793 -2.20595,-2.7962 -8.62863,-11.6647 -11.6324,-13.0661 -5.9882,-2.7896 -7.11039,4.2105 -12.13176,5.6053 -0.31809,-1.5635 -0.64246,-3.1205 -0.96682,-4.6775"
-              style={{
-                fill: "#005500",
-                fillOpacity: 0.98823529,
-                stroke: "#ffffff",
-                strokeWidth: 0.7,
-                strokeMiterlimit: 3.98288536,
-                strokeDasharray: "none",
-                strokeOpacity: 1,
-              }}
-              id="path5597"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 392.18039,111.77972 c 1.86196,-7.5711 -0.55142,-18.411898 6.55896,-23.822598 4.07418,-3.101 15.14862,0.3958 15.88811,-6.5395 -9.60798,5.6962 -1.93964,-11.4247 -0.72641,-14.5322 2.00451,-5.1382 1.94618,-7.9344 7.53206,-10.4126 4.70355,-2.089 13.7732,-6.8769 12.33291,-13.358 8.01234,4.5024 1.27156,26.0218 15.41465,20.2219 8.27155,-3.3931 8.47925,0.4801 15.67413,3.2438 10.15287,3.899 1.02489,-8.4599 0.78502,-9.2579 -1.97889,-6.4876 1.79054,-11.1976 0.96655,-17.1856 -0.83054,-5.9946 -5.74151,-11.6712 -9.27081,-16.3099 -3.97687,-5.2355 -7.49962,-10.2699 -7.92784,-17.1207996 -0.0981,-1.5441 4.88509,-18.4702004 7.66835,-15.0513 2.80261,3.4385 2.1994,5.4951 8.41438,3.8926 9.58863,-2.4783 10.60724,4.4246 16.10889,9.926 5.8192,5.8258996 12.51445,11.3014996 18.66483,16.8353996 4.71636,4.2364 4.1584,13.7278 10.56827,15.6676 2.68595,-2.874 3.31505,-6.1438 1.88131,-9.8093 1.49861,5.8064 8.18105,4.5738 9.59517,11.1198 1.42093,6.52 -2.6208,14.156 -0.66181,21.6946 5.73524,22.0254 28.89592,23.018 45.23814,33.0479 4.05482,-5.4756 2.56274,-10.5878 9.79633,-13.5462 8.64798,-3.5357 15.06439,-0.2205 22.11644,4.8592 4.08699,2.9454 10.09453,6.2346 10.92507,11.7491 0.73949,4.865698 -7.92784,4.755398 -9.16696,9.549798 4.32712,1.7646 12.346,0.4736 14.03269,6.1502 1.29091,4.3467 4.55418,4.191 7.8954,5.9946 14.12346,7.5905 8.93336,8.9529 1.88131,19.6185 9.60172,2.7508 -8.95272,14.1495 -15.80361,13.1115 -1.66734,-0.253 -18.51573,-5.4042 -19.68344,-1.9009 -1.129,3.3606 5.98792,9.1216 4.67084,14.954 -5.49483,0.052 -10.96404,6.1567 -3.20466,8.5506 12.98164,4.0159 26.95601,16.21908 38.39351,24.22479 4.50893,3.15947 26.13856,3.58114 25.5547,11.41169 -0.45411,6.14378 -4.45687,7.92138 0.90849,13.78618 3.94443,4.32074 10.6266,11.07436 15.36259,11.18464 4.86573,0.11039 7.90848,8.60908 9.58863,12.07994 3.84058,7.94732 9.61453,13.8251 15.6417,20.11808 3.46442,3.6201 16.00477,20.76686 21.27936,19.20337 5.13177,-1.51162 9.25773,15.3302 14.31808,17.49709 5.97511,2.5561 3.59417,14.85014 7.40231,19.83262 1.25875,1.64785 4.28842,8.01868 6.5527,8.37551 2.74401,0.43467 5.69599,-2.48476 8.38848,-0.92125 5.45586,3.17244 3.62007,8.42092 10.64595,10.06229 6.62384,1.53756 13.83181,11.44416 21.14335,14.22733 -2.45235,5.57936 -9.90673,8.10951 -12.02163,13.88349 -2.29672,6.27352 14.85014,14.32465 17.33493,18.88545 3.77571,6.95471 11.89162,6.92876 7.83707,17.02349 -1.20669,2.99728 -4.56099,6.26055 -4.6583,9.64058 -0.0845,3.15298 3.41918,5.32633 3.3802,9.05671 -0.13084,10.58778 -2.62108,12.23563 -11.17829,14.26626 -7.35679,1.74519 -9.30952,11.48308 -19.42395,11.63229 -9.75081,0.14937 -11.63867,-12.7287 -17.72417,-13.57207 -4.0352,-0.55793 -2.00451,4.95653 -4.67738,6.07888 -2.66006,1.11587 -5.71562,1.3689 -8.49234,2.10199 -3.25018,0.84989 -12.65728,5.57287 -10.43198,9.65356 2.95825,5.4042 10.20493,11.51552 4.82021,17.6593 -6.58486,7.50616 -21.96053,-13.43587 -32.75585,-9.47841 -6.07242,2.22525 -15.68722,-8.58959 -9.79006,-13.54614 4.41162,-3.69794 12.5537,-7.13636 11.82047,-14.02621 -0.67462,-6.40327 1.82298,-15.8687 0.44784,-21.31831 -0.9213,-3.62655 -7.0068,-4.05474 -7.0068,-8.17438 0,-3.78876 3.23737,-7.22721 4.67765,-10.51642 -1.82952,0.88232 -3.45134,0.77852 -4.87227,-0.30491 -0.94066,-5.24199 11.21727,-12.0994 3.27635,-20.15052 -4.30123,-4.3597 -21.34424,-9.52381 -10.53583,-16.90672 -18.00983,-7.06503 -36.43452,-19.1774 -39.2892,-40.52164 -2.21875,-16.57584 3.43199,-39.79501 -13.60447,-49.89623 -17.36737,-10.30234 -23.27107,-35.30559 -43.42797,-39.60688 -2.60145,-0.55793 -37.15466,-10.69807 -25.58059,2.0436 -8.129,1.13533 -5.72216,12.70275 -9.34223,16.34879 -1.42093,1.44023 -9.56274,1.59595 -12.80665,4.51536 -5.02137,4.52837 -6.96101,-0.3568 -10.53583,-3.01023 -5.96203,-4.41808 -15.71311,-10.7954 -22.80387,-13.02065 -7.21424,-2.26417 -7.47373,12.43025 -16.01159,7.31155 -0.62283,12.83897 -16.03093,21.77891 -23.47849,30.67343 -3.43853,4.10665 -8.31081,1.83599 -12.04099,-2.05009 -4.14559,-4.32723 -0.83054,-12.30051 1.03796,-16.84834 2.91301,-7.09094 3.8016,-11.8788 2.70531,-19.57312 -0.65527,-4.59324 -5.18356,-12.48218 -4.61933,-16.49151 1.16799,-2.32255 2.64071,-4.44403 4.42471,-6.35137 2.3744,-3.97692 -0.0845,-4.10018 0.18808,-8.13546 0.59694,-8.83613 3.52929,-13.94189 -0.92758,-22.3953 -6.28012,-11.9242 0.19353,-30.0506 -19.96255,-24.5101 -2.40683,-0.014 -4.19738,0.9342 -5.39126,2.861 -1.96553,4.5283 -2.52349,0.6033 -6.35126,1.3624 -6.69525,1.3364 -9.77044,-1.3559 -13.0596,-7.7981 -4.68392,-9.167 7.60347,-20.2738 2.23839,-27.0793 -2.20595,-2.7962 -8.62863,-11.6647 -11.6324,-13.0661 -5.9882,-2.7896 -7.11039,4.2105 -12.13176,5.6053 -0.31809,-1.5635 -0.64246,-3.1205 -0.96682,-4.6775"
-              id="la"
-              data-inkscape-connector-curvature="0"
-            />
-          </g>
-          <g
-            transform="matrix(6.4876122,0,0,6.4876122,-12731.067,-3673.6483)"
-            style={{
-              fill: clickedCountryId === "Timor-Leste" ? "#FF0000" : "#005500",
-              fillOpacity: 1,
-              stroke: "#ffffff",
-              strokeWidth: 0.5257113,
-              strokeMiterlimit: 3.98288536,
-              strokeDasharray: "none",
-              strokeOpacity: 1,
-            }}
-            id="Timor-Leste"
-            className="country"
-          >
-            <g transform="translate(-29.90172,-45.07447)" id="tl-">
-              <path
-                d="m 2252.9175,884.3125 c 0.055,2.702 0.6825,1.5735 2.3395,0.5695 1.584,-0.96 4.035,-1.54 5.794,-2.223 2.832,-1.1 6.902,-2.155 8.884,-4.712 1.141,-1.471 -1.928,-1.171 -2.583,-0.978 -1.78,0.526 -3.354,0.76 -5.177,1.071 -1.953,0.333 -3.761,0.296 -5.768,0.722 -0.756,0.161 -1.779,0.065 -2.417,0.58 -0.675,0.544 -1.0558,1.8335 -1.7598,2.3605 1.674,0.811 1.5113,1.266 0.6873,2.61 z"
-                id="path2786"
-                data-inkscape-connector-curvature="0"
-              />
-              <path
-                d="m 2243.958,886.033 c 1.332,-0.989 2.0745,-1.35622 3.7015,-1.66322 0.1631,1.156 0.2216,2.41032 -0.4822,3.00559 -0.1493,0.12632 -0.7474,-0.69457 -0.9442,-0.54136 -0.7152,0.55679 -1.4092,-0.53387 -2.2432,-0.83287"
-                id="path2788"
-                data-inkscape-connector-curvature="0"
-              />
-            </g>
-            <path
-              d="m 2230.8492,842.74381 c 7.7226,0 7.7226,-12.02609 0,-12.02609 -7.7282,0 -7.7282,12.02609 0,12.02609"
-              id="tl."
-              data-inkscape-connector-curvature="0"
-            />
-          </g>
-          <g
-            transform="matrix(6.4876122,0,0,6.4876122,-12731.067,-3673.6483)"
-            style={{
-              fill: clickedCountryId === "Brunei" ? "#FF0000" : "#005500",
-              fillOpacity: 0.98823529,
-              stroke: "#ffffff",
-              strokeWidth: 0.5257113,
-              strokeMiterlimit: 3.98288536,
-              strokeDasharray: "none",
-              strokeOpacity: 1,
-            }}
-            className="country"
-            id="Brunei"
-          >
-            <g transform="translate(-29.90172,-45.07447)" id="bn-">
-              <path
-                d="m 2171.518,764.928 c 2.897,-0.074 3.899,-2.64 6.333,-3.582 1.083,-0.419 0.602,1.018 0.14,1.391 -1.072,0.865 -1.277,1.026 -1.092,2.528 0.015,0.121 0.701,5.564 -1.026,4.357 -0.653,-0.456 -2.114,-1.464 -2.554,-2.173 -0.637,-1.021 -0.343,-2.209 -1.801,-2.521"
-                id="path6824"
-                data-inkscape-connector-curvature="0"
-              />
-              <path
-                d="m 2178.508,763.704 c -0.22,-1.323 0.492,-0.438 0.86,-1.512 1.05,1.343 1.499,3.288 1.37,4.968 -1.662,-0.373 -2.141,-1.943 -2.23,-3.456"
-                id="path6826"
-                data-inkscape-connector-curvature="0"
-              />
-            </g>
-            <path
-              d="m 2145.6944,727.00262 c 7.7226,0 7.7226,-12.02609 0,-12.02609 -7.7282,0 -7.7282,12.02609 0,12.02609"
-              style={{
-                opacity: 0,
-                fill: "#005500",
-                fillOpacity: 0.98823529,
-                stroke: "#ffffff",
-                strokeWidth: 0.5257113,
-                strokeMiterlimit: 3.98288536,
-                strokeDasharray: "none",
-                strokeOpacity: 1,
-              }}
-              id="bn."
-              data-inkscape-connector-curvature="0"
-            />
-          </g>
-          <g
-            transform="matrix(6.4876122,0,0,6.4876122,-12731.067,-3673.6483)"
-            style={{
-              fill: clickedCountryId === "Singapore" ? "#FF0000" : "#005500",
-              fillOpacity: 1,
-              stroke: "#ffffff",
-              strokeWidth: 0.5257113,
-              strokeMiterlimit: 3.98288536,
-              strokeDasharray: "none",
-              strokeOpacity: 1,
-            }}
-            className="country"
-            id="Singapore"
-          >
-            <path
-              d="m 2063.4263,747.21355 c -0.959,0.252 0.509,2.079 1.8,0.359 -0.569,-0.271 -1.169,-0.391 -1.8,-0.359"
-              id="sg-"
-              data-inkscape-connector-curvature="0"
-            />
-            <path
-              d="m 2065.0579,754.1131 c 7.7226,0 7.7226,-12.02609 0,-12.02609 -7.7282,0 -7.7282,12.02609 0,12.02609"
-              id="sg."
-              data-inkscape-connector-curvature="0"
-            />
-          </g>
-        </svg>
-      </div>
-    </>
-  );
-};
-
-export default Map2;
